Split the steering evaluation into modules
evaluation.py was 2510 lines holding the harness, every metric, 27
scenario definitions, the report renderers and the CLI, with the chart
traces built inside the metric computation.

Separate them: eval_spec, eval_harness, eval_metrics, eval_scenarios and
eval_compare, leaving evaluation.py as the CLI. Names the CI job and the
tests import stay available from evaluation.

Along the way, fold together the pairs that had been copied — the two
trace replayers, the two solar curves, the two Wh integrators and the
three relative-delta loops — and drop the unreachable
SimulationRunner.from_config_file. Metric values are unchanged: all 15
metrics over all 33 scenarios match the previous run exactly.
