+++ src/astrameter/simulator/evaluation.py
@@ -1,2339 +1,88 @@
-"""Steering-quality evaluation harness for the active-control loop.
+"""Steering-quality evaluation for the active-control loop.
 
-Wires the full closed loop — :class:`~astrameter.ct002.ct002.CT002` (active
-control) → :class:`~astrameter.simulator.battery.BatterySimulator` (real
-firmware steering laws) → :class:`~astrameter.simulator.load_model.LoadModel`
-→ :class:`~astrameter.simulator.powermeter_sim.PowermeterSimulator` — under a
-mock clock, so hours of simulated household activity (load spikes, solar,
-single / multiple / mixed batteries) run in seconds of wall time.
+Runs CT002 against the simulated battery plant under a mock clock, so hours of
+household activity take seconds. Each scenario (``eval_scenarios``) yields
+reaction / oscillation / energy / cost metrics (``eval_metrics``), averaged
+over several seeds run in parallel; ``--compare`` renders the base-vs-head
+Markdown CI posts on PRs (``eval_compare``), ``--html`` the interactive report
+(``eval_report``)::
 
-Each scenario produces metrics answering three questions (issue #458):
-
-* **Reaction** — how fast does the loop settle after a load/solar step?
-* **Oscillation** — how much does it overshoot and hunt around the null?
-* **Energy** — how many Wh leak to/from the grid that a battery with
-  headroom could have covered?
-
-Each scenario is run over several seeds (``--seeds``, default 5) **in parallel
-across CPU cores** and every metric is reported as the mean over those seeds,
-so the figures reflect the seed-averaged signal rather than one noisy draw.
-Process isolation (not asyncio) is what parallelizes the CPU-bound sim and, as
-a bonus, gives each seed its own global ``random`` state so runs stay
-deterministic. Per-scenario rows are then rolled up into a single cross-scenario
-``AGGREGATE`` so an across-the-board win or regression is visible at a glance.
-
-Run the suite (from the repo root, with dev deps)::
-
-    uv run python -m astrameter.simulator.evaluation
+    uv run python -m astrameter.simulator.evaluation --list
     uv run python -m astrameter.simulator.evaluation --scenario two_venus/fair \\
         --set balance_deadband=25 --seeds 10 --json head.json
     uv run python -m astrameter.simulator.evaluation --compare base.json \\
         --input head.json
-
-``--compare`` renders a Markdown before/after table — leading with the
-seed-averaged ``AGGREGATE`` roll-up — and CI runs the suite on the PR base and
-head and posts that comparison as a sticky PR comment (see
-``.github/workflows/ci.yml``, job ``steering-eval``). The interactive
-grid-power charts live in the self-contained HTML report CI uploads as the
-``steering-eval`` artifact.
-
-The roll-up leads with **two** one-line verdicts: an equal-weighted ``Overall``
-(how many metrics moved each way, mean relative change) and a ``Priority``
-verdict that weights metrics by real-world value — import-heavy self-consumption
-energy, do-no-harm overshoot/hunting guardrails, and cycle-life battery travel
-(see :data:`_METRIC_WEIGHTS`) — and hard-flags any do-no-harm guardrail
-regression (:data:`_GUARDRAIL_METRICS`: overshoot/hunting plus avoidable grid
-import — the retail-tariff money metric). The flat mean answers "did most numbers
-move down?"; the priority verdict answers "did it get better *where it matters*,
-and did it break a guardrail?".
 """
 
 from __future__ import annotations
 
 import argparse
-import asyncio
-import errno
-import functools
-import itertools
 import json
-import math
-import os
-import random
-import socket
 import sys
-from collections.abc import Callable, Sequence
-from concurrent.futures import ProcessPoolExecutor, as_completed
-from dataclasses import dataclass, field, replace
-from pathlib import Path
 
-from astrameter.ct002.balancer import device_capabilities
-from astrameter.ct002.ct002 import CT002
-
-from .battery import BatterySimulator
-from .load_model import Load, LoadModel, load_net_trace, load_power_trace
-from .powermeter_sim import PowermeterSimulator
-
-# Mock-time epoch all scenarios start from (any fixed value works; using a
-# constant keeps runs bit-for-bit reproducible across machines).
-_EPOCH = 1_750_000_000.0
-
-# |grid| below this counts as "settled" (just above the battery's own
-# ±20 W deadband, matching the main e2e convergence assertion).
-SETTLE_BAND_W = 25.0
-# The grid must stay inside SETTLE_BAND_W for this long to count as settled.
-SETTLE_HOLD_S = 10.0
-# Settling/overshoot are measured in a window after each labeled event,
-# truncated by the next labeled event.
-EVENT_WINDOW_S = 600.0
-# Oscillation counting uses the battery's deadband as hysteresis band.
-OSC_BAND_W = 20.0
-# Samples within this long after a labeled event are excluded from the
-# steady-state RMS (they're legitimate transients, not hunting).
-STEADY_EXCLUDE_S = 120.0
-# Headroom margin when deciding whether grid exchange was "avoidable".
-HEADROOM_MARGIN_W = 5.0
-SOC_EMPTY = 0.02
-SOC_FULL = 0.98
-
-# --- Cost / oracle metric (the money north-star) ---------------------------
-# Flat (time-invariant) EU-typical tariffs used to price the residual grid
-# exchange in eurocents.  Import is paid at the retail tariff, export earns the
-# (much lower) feed-in tariff.  These are the only value judgment in the cost
-# metric; the asymmetry (retail >> feed-in) is what makes a controller that
-# silently exports stored energy show up as money lost.  Costs are reported in
-# *eurocents* (not euros) and the cost keys are rounded to 2 dp everywhere
-# (per-scenario rows, seed-merge and the aggregate — see ``_metric_ndp``);
-# regret clusters at fractions of a cent, so the usual 1-dp metric rounding
-# would quantize the guardrail's 5% test into noise.
-RETAIL_CT_PER_KWH = 30.0
-FEEDIN_CT_PER_KWH = 8.0
-# Number of points each scenario's grid-power trace is downsampled to for the
-# interactive charts in the HTML report. Base and head share this fixed count
-# so the two lines align by index regardless of poll cadence.
-GRAPH_POINTS = 1800
-# Resolution of the half-sine solar day curve (s).  Fine enough that each step
-# is a few watts (vs ~50 W per minute at the steepest part of the sine), so the
-# controller sees a smooth ramp rather than a minutely staircase of load steps.
-SOLAR_CURVE_STEP_S = 2
-
-
-# ---------------------------------------------------------------------------
-# Scenario definition
-# ---------------------------------------------------------------------------
-
-
-@dataclass(frozen=True)
-class BatterySpec:
-    """Static description of one simulated battery in a scenario."""
-
-    device_type: str = "HMG-50"
-    phase: str = "A"
-    max_charge_power: int = 2500
-    max_discharge_power: int = 2500
-    capacity_wh: float = 5120.0
-    initial_soc: float = 0.6
-    # How fast measured output follows the commanded setpoint (W/s). Unlike the
-    # steering laws, this figure is **not** firmware-derived: the Venus Control
-    # MCU applies level caps and derates to the setpoint but no rate limit at
-    # all (verified — see venus_integer_steering), so the slewing belongs
-    # entirely to the separate inverter sub-processor, whose images carry no
-    # anchor to read it from. Settling it needs a bench trace (command a step,
-    # log reported power). It shapes transients only, never the control law.
-    ramp_rate: float = 400.0
-    poll_interval: float = 1.0
-    startup_delay: float = 2.0
-    min_power_threshold: float = 5.0
-    max_dc_input: int = 0
-    # Net output (W) already flowing at t=0 (positive = discharge, negative =
-    # charge). Default 0 cold-starts from rest; non-zero models a system already
-    # in motion, which the firmware input deadband treats differently from rest.
-    initial_power: float = 0.0
-
-    @property
-    def ac_chargeable(self) -> bool:
-        return device_capabilities(self.device_type).has_ac_input
-
-
-@dataclass(frozen=True)
-class Event:
-    """A scheduled world mutation.
-
-    A non-empty *label* marks a step disturbance whose settling/overshoot is
-    measured; unlabeled events (e.g. the per-minute solar curve) only mutate
-    the world.
-    """
-
-    at: float
-    apply: Callable[[EvalWorld], None]
-    label: str = ""
-
-
-@dataclass
-class Scenario:
-    name: str
-    description: str
-    batteries: list[BatterySpec]
-    duration_s: float
-    build_events: Callable[[random.Random], list[Event]]
-    base_load: list[float] = field(default_factory=lambda: [300.0, 0.0, 0.0])
-    base_noise: float = 10.0
-    loads: list[Load] = field(default_factory=list)
-    ct_kwargs: dict[str, float] = field(default_factory=dict)
-    # Real grid meters report with latency; the controller acts on a reading
-    # refreshed at this cadence (matching a typical ~1 s powermeter poll /
-    # THROTTLE_INTERVAL) while the metrics see the true instantaneous grid.
-    # Slow sources (some HTTP/MQTT meters) only emit a fresh point every ~10 s.
-    meter_interval_s: float = 1.0
-    # Transport/measurement delay on top of the refresh interval: the value the
-    # controller reads reflects the true grid as it was this many seconds ago
-    # (a P1 dongle / HA push sensor measures, then takes time to arrive). Acting
-    # on a stale reading is a classic driver of sustained oscillation, so this
-    # is what reproduces a loop that hunts instead of settling. The default is a
-    # realistic slight delay for a common ~1 s meter (no real meter is delay-
-    # free); set 0.0 only to isolate controller behaviour from meter latency.
-    meter_latency_s: float = 0.5
-
-
-@dataclass
-class EvalWorld:
-    """Mutable world handle passed to scenario events."""
-
-    load_model: LoadModel
-    batteries: list[BatterySimulator]
-    # Solar is curve x factor so labeled transients (cloud dips) compose with
-    # the unlabeled day curve instead of being overwritten by its next tick.
-    solar_curve_w: float = 0.0
-    solar_factor: float = 1.0
-
-    def set_load(self, name: str, active: bool) -> None:
-        for ld in self.load_model.loads:
-            if ld.name == name:
-                ld.active = active
-                return
-        raise KeyError(f"no load named {name!r}")
-
-    def set_base(self, watts: float, phase_index: int = 0) -> None:
-        """Set the base (whole-house) load on a phase — used to replay a
-        recorded power trace as the scenario's background consumption."""
-        self.load_model.base_load[phase_index] = watts
-
-    def set_solar_curve(self, watts: float) -> None:
-        self.solar_curve_w = watts
-        self._apply_solar()
-
-    def set_solar_factor(self, factor: float) -> None:
-        self.solar_factor = factor
-        self._apply_solar()
-
-    def _apply_solar(self) -> None:
-        self.load_model.set_solar(self.solar_curve_w * self.solar_factor)
-
-    def set_dc_input(self, battery_index: int, watts: float) -> None:
-        self.batteries[battery_index].dc_input_power = watts
-
-
-class _EvalClock:
-    """Monotonic settable mock clock (same shape as the e2e HarnessClock)."""
-
-    def __init__(self, start: float) -> None:
-        self._now = start
-
-    def __call__(self) -> float:
-        return self._now
-
-    def set(self, value: float) -> None:
-        if value > self._now:
-            self._now = value
-
-
-# ---------------------------------------------------------------------------
-# Scenario runner
-# ---------------------------------------------------------------------------
-
-
-@dataclass(frozen=True)
-class _Sample:
-    t: float  # seconds since scenario start
-    # Instantaneous *true* grid total W at this tick (sum of true_grid phases) --
-    # the physical ground truth, NOT the delayed value the controller reads from
-    # the meter cache. Recorded in before_send for the metrics/plots.
-    grid: float
-    # Raw whole-house consumption (load + noise - solar) at this instant, taken
-    # straight from the load model — the original source data, not derived from
-    # grid+battery, so the plotted consumption can never be polluted by control-
-    # loop oscillation (which lives in `grid`).
-    consumption: float
-    powers: tuple[float, ...]
-    socs: tuple[float, ...]
-    # Total DC-input power (W) across the fleet at this tick — PV wired straight
-    # into a battery's DC side (B2500, hybrid Venus), energy that never crosses
-    # the house meter so it is *not* in ``consumption``. The oracle needs it to
-    # know that free energy is arriving; 0 for pure-AC fleets.
-    dc_input: float = 0.0
-
-
-def _reserve_udp_port() -> socket.socket:
-    """Claim a free UDP port and *hold* it, returning the holding socket.
-
-    Asking the OS for port 0 and closing the socket straight away returns a
-    number, not a reservation: the port goes back in the pool immediately. The
-    evaluation runs scenarios in parallel across cores (see ``_run_tasks``), so
-    two workers were routinely handed the same one, and the loser died partway
-    through a CI run with ``OSError: [Errno 98] Address already in use``.
-
-    While this socket is open the OS will not hand that port to anyone else.
-    The caller must close it before the real listener binds, because that
-    listener binds the wildcard address, which conflicts with this loopback
-    bind on the same port — so keep the hold as short as possible and see
-    ``_start_ct002`` for what covers the remaining gap.
-    """
-    sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    try:
-        sock.bind(("127.0.0.1", 0))
-    except BaseException:
-        sock.close()
-        raise
-    return sock
-
-
-async def _start_ct002(
-    ct002: CT002, batteries: list[BatterySimulator], attempts: int = 10
-) -> None:
-    """Bind the CT listener on a free port and point *batteries* at it.
-
-    The port is picked here rather than at the top of the set-up because a
-    reservation cannot be handed to the listener atomically: ``CT002`` binds
-    the wildcard address, which conflicts with the loopback reservation, so the
-    reservation has to be released a moment before the real bind. Choosing at
-    bind time narrows that window to a couple of syscalls instead of the whole
-    scenario set-up, and losing it anyway costs one retry on a fresh port
-    rather than aborting the entire run.
-
-    The batteries are re-pointed on every attempt; they only read ``ct_port``
-    when they first send, which is after this returns.
-    """
-    for attempt in range(attempts):
-        with _reserve_udp_port() as reservation:
-            ct002.udp_port = int(reservation.getsockname()[1])
-        for battery in batteries:
-            battery.ct_port = ct002.udp_port
-        try:
-            await ct002.start()
-        except OSError as exc:
-            if exc.errno != errno.EADDRINUSE or attempt == attempts - 1:
-                raise
-        else:
-            return
-
-
-async def run_scenario(
-    scenario: Scenario,
-    seed: int = 1,
-    overrides: dict[str, float] | None = None,
-) -> dict:
-    """Run *scenario* deterministically and return its metrics dict."""
-    # The LoadModel draws noise from the global ``random``; seed it so each
-    # run is reproducible.  Event schedules use an independent stream so
-    # adding noise samples never shifts the scripted timeline.
-    random.seed(seed)
-    rng = random.Random(seed + 1)
-    events = sorted(scenario.build_events(rng), key=lambda e: e.at)
-
-    clock = _EvalClock(_EPOCH)
-    load_model = LoadModel(
-        base_load=list(scenario.base_load),
-        base_noise=scenario.base_noise,
-        loads=[Load(ld.name, ld.power, ld.phase) for ld in scenario.loads],
-    )
-    ct_mac = "112233445566"
-    batteries = [
-        BatterySimulator(
-            mac=f"02B250{i + 1:06X}",
-            phase=spec.phase,
-            ct_mac=ct_mac,
-            ct_host="127.0.0.1",
-            # Real port assigned by _start_ct002 once the CT has bound it.
-            ct_port=0,
-            meter_dev_type=spec.device_type,
-            max_charge_power=spec.max_charge_power,
-            max_discharge_power=spec.max_discharge_power,
-            capacity_wh=spec.capacity_wh,
-            initial_soc=spec.initial_soc,
-            ramp_rate=spec.ramp_rate,
-            poll_interval=spec.poll_interval,
-            min_power_threshold=spec.min_power_threshold,
-            startup_delay=spec.startup_delay,
-            max_dc_input=spec.max_dc_input,
-            initial_power=spec.initial_power,
-        )
-        for i, spec in enumerate(scenario.batteries)
-    ]
-    powermeter = PowermeterSimulator(batteries=batteries, load_model=load_model, port=0)
-    world = EvalWorld(load_model=load_model, batteries=batteries)
-
-    ct_kwargs: dict[str, float] = dict(scenario.ct_kwargs)
-    ct_kwargs.update(overrides or {})
-    ct002 = CT002(
-        udp_port=0,  # assigned by _start_ct002 below
-        ct_mac=ct_mac,
-        active_control=True,
-        clock=clock,
-        consumer_ttl=10_000_000,  # mock time spans hours; never evict
-        dedupe_time_window=0.0,
-        **ct_kwargs,
-    )
-
-    samples: list[_Sample] = []
-    # The controller reads the meter at the meter's own cadence (stale in
-    # between, like a real powermeter poll); metrics record the true grid.
-    # ``grid_history`` keeps recent true readings so a refresh can serve the
-    # value as it was ``meter_latency_s`` ago (transport/measurement delay).
-    meter_cache: dict[str, float] = {}
-    meter_read_at = [-math.inf]
-    grid_history: list[tuple[float, dict[str, float]]] = []
-
-    async def before_send(_addr, _fields=None, _consumer_id=None):
-        now = clock() - _EPOCH
-        # Draw the house load once; derive both the true grid and the raw
-        # consumption from that same sample (get_grid_contribution re-draws
-        # noise each call, so a second call would decorrelate them).
-        contribution = load_model.get_grid_contribution()
-        true_grid = powermeter.compute_grid_from(contribution)
-        grid_history.append((now, true_grid))
-        # Drop history older than what the delayed read can still need.
-        horizon = now - scenario.meter_latency_s - scenario.meter_interval_s - 1.0
-        while len(grid_history) > 1 and grid_history[0][0] < horizon:
-            grid_history.pop(0)
-        if now - meter_read_at[0] >= scenario.meter_interval_s:
-            # Serve the reading as it was meter_latency_s ago (zero-order hold
-            # on the history: the most recent sample at or before target_t).
-            target_t = now - scenario.meter_latency_s
-            delayed = grid_history[0][1]
-            for ht, hg in grid_history:
-                if ht <= target_t:
-                    delayed = hg
-                else:
-                    break
-            meter_cache.clear()
-            meter_cache.update(delayed)
-            meter_read_at[0] = now
-        samples.append(
-            _Sample(
-                t=now,
-                grid=true_grid["phase_a"] + true_grid["phase_b"] + true_grid["phase_c"],
-                consumption=contribution[0] + contribution[1] + contribution[2],
-                powers=tuple(b.current_power for b in batteries),
-                socs=tuple(b.soc for b in batteries),
-                dc_input=sum(b.dc_input_power for b in batteries),
-            )
-        )
-        return [
-            meter_cache["phase_a"],
-            meter_cache["phase_b"],
-            meter_cache["phase_c"],
-        ]
-
-    ct002.before_send = before_send
-    await _start_ct002(ct002, batteries)
-    try:
-        # Event-driven schedule: each battery polls on its own cadence
-        # (staggered starts), scripted events fire in between.
-        next_poll = [0.5 + i * 0.131 for i in range(len(batteries))]
-        marks: list[tuple[float, str]] = []
-        event_idx = 0
-        while True:
-            i = min(range(len(batteries)), key=lambda k: next_poll[k])
-            t_next = next_poll[i]
-            if t_next > scenario.duration_s:
-                break
-            while event_idx < len(events) and events[event_idx].at <= t_next:
-                ev = events[event_idx]
-                clock.set(_EPOCH + ev.at)
-                ev.apply(world)
-                if ev.label:
-                    marks.append((ev.at, ev.label))
-                event_idx += 1
-            clock.set(_EPOCH + t_next)
-            await batteries[i].step(batteries[i].poll_interval)
-            next_poll[i] = t_next + batteries[i].poll_interval
-    finally:
-        await ct002.stop()
-
-    return _compute_metrics(scenario, seed, samples, marks)
-
-
-# ---------------------------------------------------------------------------
-# Metrics
-# ---------------------------------------------------------------------------
-
-
-def _percentile(values: Sequence[float], pct: float) -> float:
-    if not values:
-        return 0.0
-    ordered = sorted(values)
-    k = (len(ordered) - 1) * pct
-    lo = math.floor(k)
-    hi = math.ceil(k)
-    if lo == hi:
-        return ordered[lo]
-    return ordered[lo] + (ordered[hi] - ordered[lo]) * (k - lo)
-
-
-def _settle_time(samples: list[_Sample], start: float, end: float) -> float | None:
-    """Seconds from *start* until |grid| stays inside SETTLE_BAND_W for
-    SETTLE_HOLD_S, or ``None`` if it never settles inside the window."""
-    window = [s for s in samples if start <= s.t <= end]
-    candidate: float | None = None
-    for s in window:
-        if abs(s.grid) < SETTLE_BAND_W:
-            if candidate is None:
-                candidate = s.t
-            if s.t - candidate >= SETTLE_HOLD_S:
-                return candidate - start
-        else:
-            candidate = None
-    # A quiet tail shorter than the hold still counts when the window ends.
-    if candidate is not None and window and window[-1].t - candidate >= SETTLE_HOLD_S:
-        return candidate - start
-    return None
-
-
-def _downsample_series(
-    samples: list[_Sample],
-    duration_s: float,
-    pick: Callable[[_Sample], float],
-    n: int = GRAPH_POINTS,
-) -> list[float]:
-    """Bucket a per-sample value into *n* evenly spaced means over the run.
-
-    *pick* selects the value from each sample (grid, a battery's power, ...).
-    Empty buckets carry the previous value forward so the chart has no gaps;
-    the fixed length lets traces from different runs overlay by index.
-    """
-    if not samples or duration_s <= 0 or n <= 0:
-        return []
-    buckets: list[list[float]] = [[] for _ in range(n)]
-    for s in samples:
-        idx = min(int(s.t / duration_s * n), n - 1)
-        buckets[idx].append(pick(s))
-    out: list[float] = []
-    last = 0.0
-    for bucket in buckets:
-        if bucket:
-            last = sum(bucket) / len(bucket)
-        out.append(round(last, 1))
-    return out
-
-
-def _battery_pick(i: int) -> Callable[[_Sample], float]:
-    """Return a picker for battery *i*'s output (a typed closure, so the
-    per-battery downsampling avoids an inline lambda mypy can't infer)."""
-    return lambda s: s.powers[i]
-
-
-def _grid_cost_ct(import_wh: float, export_wh: float) -> float:
-    """Electricity bill (eurocents) for a residual grid exchange: import paid at
-    the retail tariff, export credited at the (lower) feed-in tariff.  Can go
-    negative when feed-in credit exceeds import cost."""
-    return (RETAIL_CT_PER_KWH * import_wh - FEEDIN_CT_PER_KWH * export_wh) / 1000.0
-
-
-def _oracle_cost_ct(scenario: Scenario, samples: list[_Sample]) -> float:
-    """Cost (eurocents) of the *perfect-foresight* battery dispatch — the
-    achievable floor this scenario's controller is benchmarked against.
-
-    The oracle runs a single ideal aggregate battery (summed capacity / charge /
-    discharge power across the fleet, lossless to match the simulator's lossless
-    plant) with perfect foresight and instantaneous response.  Each step it picks
-    the battery's net AC output ``p`` to zero the grid if it can, otherwise to
-    store / shed the residual optimally.  Two energy streams feed it:
-      * the **household net load** ``consumption`` (load + noise - AC solar) — the
-        deficit/surplus the controller fights to zero out; and
-      * **DC-input solar** ``dc_input`` — PV wired into a battery's DC side, which
-        never crosses the house meter (so it is *not* in ``consumption``).  The
-        cells gain ``dc - p`` each step: DC solar first offsets the grid, the rest
-        charges the pack, and at a full pack it passes straight through to export
-        (matching the real B2500 / hybrid-Venus plant).
-
-    Under a flat tariff with a lossless battery this greedy dispatch *is* the
-    optimum: every stored Wh is equally valuable, foresight cannot conjure energy
-    a causal controller lacks, and ``retail >= feed-in`` makes storing surplus
-    always worth more than exporting it.  What it leaves on the grid is therefore
-    physically irreducible — surplus when the pack is full, deficit when it is
-    empty or power-limited — so ``actual_cost - oracle_cost`` isolates the cost
-    the *controller* could have avoided, and the oracle is a true lower bound (so
-    ``cost_regret_ct`` stays >= 0).
-
-    One approximation remains: **per-phase routing is ignored** (the fleet is one
-    battery).  A real per-phase controller can be forced to import on one phase
-    while exporting on another, which the netted aggregate avoids — so the oracle
-    is mildly *optimistic* (still a valid lower bound) in the three-phase
-    scenario."""
-    specs = scenario.batteries
-    cap_wh = sum(s.capacity_wh for s in specs)
-    max_charge = sum(s.max_charge_power for s in specs)
-    max_discharge = sum(s.max_discharge_power for s in specs)
-    energy_wh = sum(s.initial_soc * s.capacity_wh for s in specs)
-    import_wh = export_wh = 0.0
-    for prev, cur in itertools.pairwise(samples):
-        dt = min(cur.t - prev.t, 5.0)
-        if dt <= 0:
-            continue
-        h = dt / 3600.0
-        net = prev.consumption  # >0 deficit (would import), <0 surplus (export)
-        dc = prev.dc_input  # free DC-side solar entering the pack this step
-        # Feasible net AC output p (positive = to house/grid). The cells change by
-        # (dc - p) per step, bounded by stored energy (can't over-drain) and room
-        # (can't over-fill); p is also bounded by the AC charge/discharge limits.
-        hi = min(max_discharge, dc + energy_wh / h)
-        lo = max(-max_charge, dc - (cap_wh - energy_wh) / h)
-        if lo > hi:  # DC inflow exceeds what a full pack can shed via the inverter
-            lo = hi  # → curtail the excess (output at the cap)
-        p = min(max(net, lo), hi)  # zero the grid if feasible, else store/shed
-        energy_wh = max(0.0, min(cap_wh, energy_wh + (dc - p) * h))
-        grid = net - p
-        if grid > 0:
-            import_wh += grid * h
-        else:
-            export_wh += -grid * h
-    return _grid_cost_ct(import_wh, export_wh)
-
-
-def _compute_metrics(
-    scenario: Scenario,
-    seed: int,
-    samples: list[_Sample],
-    marks: list[tuple[float, str]],
-) -> dict:
-    duration_h = scenario.duration_s / 3600.0
-    specs = scenario.batteries
-
-    # --- per-event settling & overshoot ---
-    settle_times: list[float] = []
-    overshoots: list[float] = []
-    unsettled = 0
-    events_measured = 0
-    for idx, (t0, _label) in enumerate(marks):
-        t_end = min(
-            scenario.duration_s,
-            t0 + EVENT_WINDOW_S,
-            marks[idx + 1][0] if idx + 1 < len(marks) else float("inf"),
-        )
-        window = [s for s in samples if t0 <= s.t <= t_end]
-        if not window:
-            continue
-        e0 = window[0].grid
-        if abs(e0) < SETTLE_BAND_W:
-            continue  # disturbance too small to measure against the band
-        events_measured += 1
-        sign = 1.0 if e0 > 0 else -1.0
-        settle = _settle_time(samples, t0, t_end)
-        if settle is None:
-            unsettled += 1
-            settle_times.append(t_end - t0)
-        else:
-            settle_times.append(settle)
-        overshoots.append(max(0.0, max(-sign * s.grid for s in window)))
-
-    # --- oscillation: hysteresis band crossings ---
-    crossings = 0
-    state = 0
-    for s in samples:
-        if s.grid > OSC_BAND_W:
-            if state == -1:
-                crossings += 1
-            state = 1
-        elif s.grid < -OSC_BAND_W:
-            if state == 1:
-                crossings += 1
-            state = -1
-
-    # --- steady-state RMS (outside post-event transients) ---
-    def in_transient(t: float) -> bool:
-        return any(t0 <= t < t0 + STEADY_EXCLUDE_S for t0, _ in marks)
-
-    steady = [s.grid for s in samples if not in_transient(s.t)]
-    steady_rms = math.sqrt(sum(g * g for g in steady) / len(steady)) if steady else 0.0
-
-    # --- sustained oscillation amplitude ---
-    # The robust peak-to-peak grid swing (p95 - p5) over the whole run. Unlike
-    # the step-response metrics (settle/overshoot, which only fire on labelled
-    # load steps and read 0 for a continuously hunting loop), this is non-zero
-    # for *any* sustained oscillation and grades it directly: a loop that holds
-    # zero scores ~0, one that constantly swings ±X scores ~2X. Percentiles (not
-    # min/max) keep a single brief transient from dominating.
-    all_grid = [s.grid for s in samples]
-    grid_p2p = _percentile(all_grid, 0.95) - _percentile(all_grid, 0.05)
-
-    # --- energy, battery travel & whole-run L2 tracking error ---
-    # grid_rms is RMS grid power over the *entire* run, transients included
-    # (unlike steady_rms_w, which excludes post-event windows). It's the L2
-    # tracking term of the classic control objective: squaring punishes large
-    # excursions (overshoot, big swings) far harder than a small steady offset, so
-    # it reads "how cleanly did the loop hold zero?" as one number. Its effort
-    # partner is battery_travel_w_per_h; the two are kept separate rather than
-    # fused with an arbitrary trade-off constant. Both grid_rms and mean_abs are
-    # *time-weighted* (by dt) so they're true time-averages, not sample-averages
-    # skewed by uneven / staggered poll spacing (slow meters, multi-battery).
-    # --- inter-battery share imbalance (issue #523) ---
-    # Batteries sharing a phase should split the phase's net output evenly (the
-    # balancer's fair-share target with the eval's equal distribution weights);
-    # a controller that leaves them lopsided — e.g. one Venus charging at 88 W
-    # while its twin takes 890 W — is doing real harm (uneven cycle wear, a pack
-    # that saturates first). ``share_imbalance_w`` is the time-weighted mean of
-    # the per-sample watts misallocated: within each phase group of >=2 active
-    # batteries, the sum of |power_i - phase_fair_share|. It is 0 for a perfectly
-    # balanced pool and 0 by construction for any scenario with at most one
-    # battery per phase (nothing to balance).
-    phase_groups: dict[str, list[int]] = {}
-    for i, sp in enumerate(specs):
-        phase_groups.setdefault((sp.phase or "A").upper(), []).append(i)
-    balance_groups = [grp for grp in phase_groups.values() if len(grp) >= 2]
-
-    import_wh = export_wh = avoid_import_wh = avoid_export_wh = 0.0
-    travel_w = 0.0
-    grid_sq_dt = abs_grid_dt = total_dt = 0.0
-    imbalance_dt = 0.0
-    for prev, cur in itertools.pairwise(samples):
-        dt = min(cur.t - prev.t, 5.0)
-        if dt <= 0:
-            continue
-        grid_sq_dt += prev.grid * prev.grid * dt
-        abs_grid_dt += abs(prev.grid) * dt
-        total_dt += dt
-        for grp in balance_groups:
-            fair = sum(prev.powers[i] for i in grp) / len(grp)
-            imbalance_dt += sum(abs(prev.powers[i] - fair) for i in grp) * dt
-        wh = prev.grid * dt / 3600.0
-        if wh > 0:
-            import_wh += wh
-            # Import is avoidable while any battery still has discharge
-            # headroom and charge in the pack.
-            if any(
-                prev.socs[i] > SOC_EMPTY
-                and prev.powers[i] < specs[i].max_discharge_power - HEADROOM_MARGIN_W
-                for i in range(len(specs))
-            ):
-                avoid_import_wh += wh
-        else:
-            export_wh += -wh
-            # Export is avoidable while any AC-chargeable battery has charge
-            # headroom and room in the pack.
-            if any(
-                specs[i].ac_chargeable
-                and prev.socs[i] < SOC_FULL
-                and prev.powers[i] > -specs[i].max_charge_power + HEADROOM_MARGIN_W
-                for i in range(len(specs))
-            ):
-                avoid_export_wh += -wh
-        travel_w += sum(abs(cur.powers[i] - prev.powers[i]) for i in range(len(specs)))
-
-    grid_rms = math.sqrt(grid_sq_dt / total_dt) if total_dt > 0 else 0.0
-    mean_abs = abs_grid_dt / total_dt if total_dt > 0 else 0.0
-    share_imbalance = imbalance_dt / total_dt if total_dt > 0 else 0.0
-
-    # --- money: cost vs perfect-foresight oracle (the north-star metric) ---
-    # The actual electricity bill (eurocents) for the residual grid this
-    # controller left, minus what an ideal perfect-foresight battery would have
-    # paid on the same load. The difference is the money the *controller* could
-    # have saved — ungameable (both import and export cost), 0 = matched perfect
-    # foresight. The oracle is a true lower bound (it sees the same load *and* the
-    # DC-input solar — see ``_oracle_cost_ct``), so the max(0.0, …) clamp only
-    # ever absorbs tiny per-phase-routing slack in the three-phase scenario, not a
-    # real "controller beat the optimum".
-    grid_cost = _grid_cost_ct(import_wh, export_wh)
-    oracle_cost = _oracle_cost_ct(scenario, samples)
-    cost_regret = max(0.0, grid_cost - oracle_cost)
-
-    # Lowest / highest state of charge any battery reached over the run — lets a
-    # scenario verify it actually drives the pack into empty/full saturation
-    # (not reported in the metric tables; available for tests and context).
-    all_socs = [soc for s in samples for soc in s.socs]
-    soc_min = round(min(all_socs), 3) if all_socs else 0.0
-    soc_max = round(max(all_socs), 3) if all_socs else 0.0
-
-    return {
-        "scenario": scenario.name,
-        "seed": seed,
-        "duration_h": round(duration_h, 3),
-        "samples": len(samples),
-        "soc_min": soc_min,
-        "soc_max": soc_max,
-        "events_measured": events_measured,
-        "unsettled_events": unsettled,
-        "settle_mean_s": round(sum(settle_times) / len(settle_times), 1)
-        if settle_times
-        else 0.0,
-        "settle_p95_s": round(_percentile(settle_times, 0.95), 1),
-        "overshoot_mean_w": round(sum(overshoots) / len(overshoots), 1)
-        if overshoots
-        else 0.0,
-        "overshoot_max_w": round(max(overshoots), 1) if overshoots else 0.0,
-        "band_crossings_per_h": round(crossings / duration_h, 2),
-        "grid_p2p_w": round(grid_p2p, 1),
-        "grid_rms_w": round(grid_rms, 1),
-        "steady_rms_w": round(steady_rms, 1),
-        "mean_abs_grid_w": round(mean_abs, 1),
-        "share_imbalance_w": round(share_imbalance, 1),
-        "import_wh": round(import_wh, 1),
-        "export_wh": round(export_wh, 1),
-        "avoidable_import_wh": round(avoid_import_wh, 1),
-        "avoidable_export_wh": round(avoid_export_wh, 1),
-        "grid_cost_ct": round(grid_cost, 2),
-        "oracle_cost_ct": round(oracle_cost, 2),
-        "cost_regret_ct": round(cost_regret, 2),
-        "battery_travel_w_per_h": round(travel_w / duration_h, 0),
-        "grid_trace": _downsample_series(
-            samples, scenario.duration_s, lambda s: s.grid
-        ),
-        # Raw whole-house consumption, recorded straight from the load model
-        # (the original source data — not derived from grid+battery, so it can't
-        # be polluted by control-loop oscillation). It's the same scripted load
-        # in base and head, so one trace is enough; the HTML grid chart overlays
-        # it as context (grid = consumption minus battery output).
-        "consumption_trace": _downsample_series(
-            samples, scenario.duration_s, lambda s: s.consumption
-        ),
-        # Per-battery output traces (one downsampled series each) and labels,
-        # for the per-scenario battery-output chart in the HTML report.
-        "battery_labels": [
-            f"B{i + 1} {specs[i].device_type}" for i in range(len(specs))
-        ],
-        "battery_traces": [
-            _downsample_series(samples, scenario.duration_s, _battery_pick(i))
-            for i in range(len(specs))
-        ],
-    }
-
-
-# ---------------------------------------------------------------------------
-# Scenarios
-# ---------------------------------------------------------------------------
-
-_VENUS = BatterySpec()  # HMG-50 (V2-class), 1 s poll
-# Venus E (VNSE3-0), the fast-polling V3-class unit. Like every other
-# AC-coupled Venus it runs the integer integrator behind its input gate (see
-# venus_integer_steering) — no gain-scheduled ramp, so it slews far harder than
-# an HMG-50 on a big step.
-_VENUS_V3 = BatterySpec(device_type="VNSE3-0", poll_interval=0.45)
-_VENUS_V2_SLOW = BatterySpec(poll_interval=3.1)
-# Venus D (VNSD-0): the integer proportional-integrator loop rather than the
-# HMG-50 float ramp — a ±10 W input deadband with a one-shot spike filter, a
-# ±11 W setpoint park, and a unity-by-default ctrl_ratio gain.
-_VENUS_D = BatterySpec(device_type="VNSD-0")
-_B2500 = BatterySpec(
-    device_type="HMA-1",
-    max_charge_power=0,
-    max_discharge_power=800,
-    capacity_wh=2240.0,
-    max_dc_input=1000,
-    initial_soc=0.5,
+from .eval_compare import (
+    _GUARDRAIL_METRICS,
+    _METRIC_GLOSSARY,
+    _METRIC_WEIGHTS,
+    _REPORT_METRICS,
+    _aggregate,
+    _compare_aggregates,
+    _fmt_delta,
+    _guardrail_regressions,
+    _merge_seeds,
+    _metric_ndp,
+    _overall_summary,
+    _priority_summary,
+    _seed_label,
+    _seeds_caption,
+    _weighted_overall,
+    render_markdown_compare,
+    render_text,
 )
-
-# A B2500 in a same-phase pair, sized to serve a household on its own charge
-# for the run: no DC input, so the only thing deciding its output is what the
-# balancer commands.  Matches the pair in issue #600 (two HMJ-2 on phase A).
-_B2500_PAIR = BatterySpec(
-    device_type="HMJ-2",
-    phase="A",
-    max_charge_power=0,
-    max_discharge_power=800,
-    capacity_wh=5120.0,
-    initial_soc=0.8,
+from .eval_harness import _reserve_udp_port, _run_tasks, run_scenario
+from .eval_metrics import (
+    FEEDIN_CT_PER_KWH,
+    RETAIL_CT_PER_KWH,
+    _grid_cost_ct,
+    _oracle_cost_ct,
+    _Sample,
 )
+from .eval_report import GRAPH_POINTS, render_html_report
+from .eval_scenarios import build_scenarios
+from .eval_spec import BatterySpec, Event, Scenario
 
-# Efficiency-optimization mode knobs (mirrors a typical multi-battery setup).
-_EFF_MODE: dict[str, float] = {
-    "min_efficient_power": 150.0,
-    "efficiency_rotation_interval": 900.0,
-}
-
-# Efficiency knobs for the real-trace two-Venus scenario. The concentration
-# cut-in is roughly at total load = 2 x min_efficient_power (for two units), so a
-# higher floor than the default 150 W is needed to land that threshold inside the
-# recorded load's range: at 500 W the second Venus idles on the calm overnight
-# base (~66% of the run) and cuts in on the morning/cooking peaks (~33%). With
-# the default 150 W floor both units stay active the whole run and /eff would be
-# an exact copy of /fair.
-_TRACE_EFF_MODE: dict[str, float] = {**_EFF_MODE, "min_efficient_power": 500.0}
-
-
-# Typed closure factories for event actions (a plain lambda with extra
-# defaulted parameters doesn't type-check against ``Callable[[EvalWorld], None]``).
-
-
-def _set_load(name: str, active: bool) -> Callable[[EvalWorld], None]:
-    def apply(w: EvalWorld) -> None:
-        w.set_load(name, active)
-
-    return apply
-
-
-def _set_solar_curve(watts: float) -> Callable[[EvalWorld], None]:
-    def apply(w: EvalWorld) -> None:
-        w.set_solar_curve(watts)
-
-    return apply
-
-
-def _set_solar_factor(factor: float) -> Callable[[EvalWorld], None]:
-    def apply(w: EvalWorld) -> None:
-        w.set_solar_factor(factor)
-
-    return apply
-
-
-def _set_dc_input(battery_index: int, watts: float) -> Callable[[EvalWorld], None]:
-    def apply(w: EvalWorld) -> None:
-        w.set_dc_input(battery_index, watts)
-
-    return apply
-
-
-def _set_base(watts: float) -> Callable[[EvalWorld], None]:
-    def apply(w: EvalWorld) -> None:
-        w.set_base(watts)
-
-    return apply
-
-
-def _set_net(load_w: float, pv_w: float) -> Callable[[EvalWorld], None]:
-    """Set the base load and the solar (PV) curve together from one recorded
-    net-load sample, so the two stay correlated (same site, same instant)."""
-
-    def apply(w: EvalWorld) -> None:
-        w.set_base(load_w)
-        w.set_solar_curve(pv_w)
-
-    return apply
-
-
-def _household_steps(rng: random.Random, duration: float) -> list[Event]:
-    """Scripted appliance schedule: kettle spikes, oven cycling, dishwasher.
-
-    Times get a deterministic per-seed jitter so different seeds exercise
-    different alignments against the poll cadence.
-    """
-
-    def jitter(t: float, spread: float = 20.0) -> float:
-        return max(1.0, t + rng.uniform(-spread, spread))
-
-    events: list[Event] = []
-
-    def load_event(t: float, name: str, active: bool) -> None:
-        state = "on" if active else "off"
-        events.append(
-            Event(at=jitter(t), label=f"{name}_{state}", apply=_set_load(name, active))
-        )
-
-    # Kettle: two 2 kW bursts of ~3 minutes.
-    for burst, t0 in enumerate((600.0, duration * 0.7)):
-        load_event(t0, "kettle", True)
-        load_event(t0 + 180.0, "kettle", False)
-        del burst
-    # Oven: thermostat cycling between 30% and 60% of the run.  Cycles are
-    # emitted as on/off pairs so the oven never stays on past its block
-    # (an unpaired trailing "on" would stack loads beyond the battery's
-    # ceiling for the rest of the run).
-    t = duration * 0.3
-    while t + 240.0 < duration * 0.6:
-        load_event(t, "oven", True)
-        load_event(t + 240.0, "oven", False)
-        t += 240.0 + 180.0
-    # Dishwasher: one long block in the second half.
-    load_event(duration * 0.8, "dishwasher", True)
-    load_event(duration * 0.8 + 600.0, "dishwasher", False)
-    return events
-
-
-_HOUSEHOLD_LOADS = [
-    Load("kettle", 2000.0, "A"),
-    Load("oven", 1500.0, "A"),
-    Load("dishwasher", 1100.0, "A"),
+__all__ = [
+    "FEEDIN_CT_PER_KWH",
+    "GRAPH_POINTS",
+    "RETAIL_CT_PER_KWH",
+    "_GUARDRAIL_METRICS",
+    "_METRIC_GLOSSARY",
+    "_METRIC_WEIGHTS",
+    "_REPORT_METRICS",
+    "BatterySpec",
+    "Event",
+    "Scenario",
+    "_Sample",
+    "_aggregate",
+    "_compare_aggregates",
+    "_fmt_delta",
+    "_grid_cost_ct",
+    "_guardrail_regressions",
+    "_merge_seeds",
+    "_metric_ndp",
+    "_oracle_cost_ct",
+    "_overall_summary",
+    "_priority_summary",
+    "_reserve_udp_port",
+    "_run_all",
+    "_seed_label",
+    "_weighted_overall",
+    "build_scenarios",
+    "main",
+    "render_markdown_compare",
+    "render_text",
+    "run_scenario",
 ]
-
-# Same appliances spread one-per-phase, for the three-phase imbalance scenario
-# (everything else in the suite is single-phase A). Names match _household_steps
-# so the same scripted schedule drives them.
-_PHASE_IMBALANCE_LOADS = [
-    Load("kettle", 2000.0, "A"),
-    Load("oven", 1500.0, "B"),
-    Load("dishwasher", 1100.0, "C"),
-]
-
-# Washing-machine drum motor: a single ~120 W load the main-wash tumble runs,
-# briefly pauses, and restarts.  Sized (with the scenario's ~1 s meter latency)
-# to reproduce the field report in issue #473 — a steady ~500 W house whose
-# grid never holds zero, hunting on the order of the log's ±100-180 W swings.
-# Fidelity notes: the simulated battery plant limit-cycles somewhat harder and
-# faster than the real (better-damped) firmware, so the sustained swing here is
-# larger/quicker than that one log; and the modelled pause drops the full
-# running load, so the export dip is about as deep as the import spike, whereas
-# the field trace was asymmetric (shallower dip, larger spike) due to motor
-# restart inrush this single on/off load does not model.
-_WASHER_LOADS = [Load("washer_motor", 120.0, "A")]
-
-
-def _washer_cycle(rng: random.Random, duration: float) -> list[Event]:
-    """Main-wash drum tumble: the motor runs, briefly pauses, and restarts.
-
-    This reproduces the field-reported washing-machine signature (issue #473).
-    A real drum tumbles in one direction, briefly pauses, then reverses, so a
-    short pause drops the load (a brief **export dip** as the battery is still
-    discharging) and the restart re-applies it (an **import spike**), repeating
-    every ~16 s. Paired with the scenario's ~1 s meter latency, the loop acts on
-    stale readings and never fully settles between pauses, so the grid hunts
-    continuously rather than holding zero — matching the log, which never showed
-    a steady-at-zero phase.
-
-    The events are unlabelled on purpose: a continuously hunting loop never
-    holds the ±25 W band for the settle hold time, so this scenario is scored on
-    the sustained-oscillation aggregates (``grid_p2p_w``, ``band_crossings_per_h``,
-    ``steady_rms_w``, ``mean_abs_grid_w``, ``battery_travel_w_per_h``) rather than
-    per-step settling — the step-response metrics read 0 for this failure mode.
-    A balancer that damps the hunt drives those down.
-    """
-    events: list[Event] = []
-    period = 16.0
-    pause = 3.0
-    start = duration * 0.15
-    end = duration * 0.85
-    # Motor on for the whole wash block; the rhythm is the brief pauses.
-    events.append(Event(at=start, apply=_set_load("washer_motor", True)))
-    t = start + (period - pause)
-    while t + pause < end:
-        # Small deterministic per-pause jitter so the rhythm doesn't phase-lock
-        # to the 1 s meter cadence (and different seeds probe different
-        # alignments), without ever reordering the pause/restart pair.
-        j = rng.uniform(-0.5, 0.5)
-        events.append(Event(at=max(1.0, t + j), apply=_set_load("washer_motor", False)))
-        events.append(
-            Event(at=max(1.0, t + pause + j), apply=_set_load("washer_motor", True))
-        )
-        t += period
-    # Always leave the program with the motor off.
-    events.append(Event(at=end, apply=_set_load("washer_motor", False)))
-    return events
-
-
-# "Pretty noisy" house baseline: a net draw that jitters hard sample-to-sample
-# (many small switching loads — fridge, electronics, pumps, chargers — with no
-# single dominant appliance). The ±_NOISY_BASE_NOISE W uniform jitter on every
-# meter read is the only disturbance: there are no scripted load steps, so (like
-# the washer scenario) this is scored on the sustained-oscillation aggregates
-# (grid_p2p_w, band_crossings_per_h, steady_rms_w, mean_abs_grid_w,
-# battery_travel_w_per_h) rather than per-step settling — a balancer that ignores
-# the noise (spike filter / deadband / damping) holds grid swing and battery
-# travel down instead of chasing every wiggle.
-_NOISY_BASE_LOAD = [400.0, 0.0, 0.0]
-_NOISY_BASE_NOISE = 150.0
-
-
-def _no_events(_rng: random.Random) -> list[Event]:
-    """A scenario with no scripted events: only the base load and its noise (plus
-    any solar) drive the loop, so the loop is exercised purely on noise rejection
-    and scored on the sustained-oscillation aggregates (the step-response metrics
-    read 0 with no labeled events)."""
-    return []
-
-
-# Real household whole-house power trace (RAE House 1, 1 Hz, CC BY — see
-# traces/README.md). Unlike the synthetic base load + IID noise, a recorded
-# trace carries the *correlated drift* and *persistent appliance switching* of a
-# real home, so it rewards a balancer that tracks genuine load changes instead
-# of one tuned to reject white noise (which over-damps and then lags in the
-# field). Loaded lazily on first use (not at import) and cached; scenarios slice
-# a per-seed window from it.
-_TRACE_DIR = Path(__file__).parent / "traces"
-
-
-@functools.cache
-def _household_trace() -> list[tuple[float, float]]:
-    """Lazily load + cache the real household power trace (RAE House 1).
-
-    Loading on first use rather than at import keeps the module import (and so
-    ``--help`` / importing helpers for tests) side-effect-free, and turns a
-    missing trace into a clear, scenario-specific error instead of an opaque
-    import-time traceback. The vendored CSVs are dev/eval-only data and aren't
-    shipped in the wheel.
-    """
-    path = _TRACE_DIR / "rae_household.csv"
-    try:
-        return load_power_trace(path)
-    except OSError as exc:
-        raise RuntimeError(
-            f"Real-trace scenario needs the household trace at {path}, which "
-            "isn't available (dev/eval-only data, not packaged) — run from a "
-            "source checkout. See traces/README.md."
-        ) from exc
-
-
-# Small measurement noise laid over the recorded trace: a real CT/meter adds a
-# few watts of jitter on top of the true house power. The 1 Hz trace already
-# carries real second-to-second structure, so this only models meter noise (not
-# a zero-order-hold fill).
-_TRACE_METER_NOISE = 10.0
-
-
-def _trace_events(rng: random.Random, duration: float) -> list[Event]:
-    """Replay a per-seed window of the real household trace as the base load.
-
-    Each seed picks a different start offset into the recording (so seeds see
-    genuinely different load regimes — quiet stretches vs cooking spikes — not
-    just re-randomised noise), then the recorded samples drive the whole-house
-    base load over the run. The events are **unlabelled**: real load changes
-    aren't clean single steps, so (like the washer/noisy scenarios) this is
-    scored on the sustained-oscillation/energy aggregates rather than per-step
-    settling. A balancer that tracks the real load without hunting drives those
-    down.
-    """
-    trace = _household_trace()
-    t0 = trace[0][0]
-    span = trace[-1][0] - t0
-    max_off = max(0.0, span - duration)
-    start = rng.uniform(0.0, max_off)
-    events: list[Event] = []
-    for t, watts in trace:
-        rel = (t - t0) - start
-        if rel < 0.0:
-            continue
-        if rel > duration:
-            break
-        events.append(Event(at=rel, apply=_set_base(watts)))
-    # Anchor the base load from t=0 to the sample active at `start`, so the run
-    # opens on real load rather than the Scenario's placeholder base_load.
-    if not events or events[0].at > 0.0:
-        first = trace[0][1]
-        for t, watts in trace:
-            if (t - t0) <= start:
-                first = watts
-            else:
-                break
-        events.insert(0, Event(at=0.0, apply=_set_base(first)))
-    return events
-
-
-# Real prosumer net-load trace (Cyprus, 30 s, CC BY — see traces/README.md):
-# matching PV generation + load from one site, so their timing is genuinely
-# correlated. A partly-cloudy day, so the PV carries real cloud transients the
-# synthetic half-sine lacks. Scaled to a balcony-system size (keeps PV under the
-# load model's 2 kW solar clamp while preserving the cloud-dip shape).
-# Loaded lazily on first use (not at import) and cached, like the household trace.
-_PV_SCALE = 0.45
-
-
-@functools.cache
-def _net_trace() -> list[tuple[float, float, float]]:
-    """Lazily load + cache the real prosumer net-load trace (Cyprus).
-
-    Same lazy/cached rationale as :func:`_household_trace`.
-    """
-    path = _TRACE_DIR / "cyprus_netload.csv"
-    try:
-        return load_net_trace(path)
-    except OSError as exc:
-        raise RuntimeError(
-            f"PV-net scenario needs the net-load trace at {path}, which isn't "
-            "available (dev/eval-only data, not packaged) — run from a source "
-            "checkout. See traces/README.md."
-        ) from exc
-
-
-def _pv_net_events(rng: random.Random, duration: float) -> list[Event]:
-    """Replay a per-seed window of the real PV + load net-load trace.
-
-    Drives the base load and the solar (PV) curve together from the same
-    recorded sample (so they stay correlated), scaled by :data:`_PV_SCALE`. Each
-    seed slices a different offset, so seeds see different parts of the day
-    (morning ramp, cloudy midday, afternoon). Unlabelled — scored on the
-    sustained-oscillation/energy aggregates, like the load-trace scenarios.
-    """
-    trace = _net_trace()
-    t0 = trace[0][0]
-    span = trace[-1][0] - t0
-    max_off = max(0.0, span - duration)
-    start = rng.uniform(0.0, max_off)
-    events: list[Event] = []
-    for t, load_w, pv_w in trace:
-        rel = (t - t0) - start
-        if rel < 0.0:
-            continue
-        if rel > duration:
-            break
-        events.append(
-            Event(at=rel, apply=_set_net(load_w * _PV_SCALE, pv_w * _PV_SCALE))
-        )
-    # Anchor both load and PV from t=0 to the sample active at `start`.
-    if not events or events[0].at > 0.0:
-        first = trace[0]
-        for sample in trace:
-            if (sample[0] - t0) <= start:
-                first = sample
-            else:
-                break
-        events.insert(
-            0, Event(at=0.0, apply=_set_net(first[1] * _PV_SCALE, first[2] * _PV_SCALE))
-        )
-    return events
-
-
-def _solar_curve(duration: float, peak: float) -> list[Event]:
-    """Unlabeled half-sine solar day curve.
-
-    Updated every ``SOLAR_CURVE_STEP_S`` so the curve is smooth: a coarse
-    per-minute step jumps ~50 W per tick where the sine is steepest (sunrise /
-    sunset), which the controller would chase as a series of small load steps.
-    """
-    events: list[Event] = []
-    for t in range(0, int(duration), SOLAR_CURVE_STEP_S):
-        watts = peak * math.sin(math.pi * t / duration)
-        events.append(Event(at=float(t), apply=_set_solar_curve(watts)))
-    return events
-
-
-def _cloud_dips(rng: random.Random, duration: float) -> list[Event]:
-    """Labeled cloud transients: solar collapses to 20% for ~2 minutes.
-
-    Implemented as a multiplicative factor so the day curve keeps ticking
-    underneath without cancelling the dip.
-    """
-    events: list[Event] = []
-    for frac in (0.4, 0.55):
-        t0 = duration * frac + rng.uniform(-60.0, 60.0)
-        events.append(Event(at=t0, label="cloud_on", apply=_set_solar_factor(0.2)))
-        events.append(
-            Event(at=t0 + 120.0, label="cloud_off", apply=_set_solar_factor(1.0))
-        )
-    return events
-
-
-def _dc_solar_curve(duration: float, peak: float, battery_index: int) -> list[Event]:
-    """Unlabeled DC-input solar curve for a B2500-style battery (smooth — see
-    :func:`_solar_curve`)."""
-    events: list[Event] = []
-    for t in range(0, int(duration), SOLAR_CURVE_STEP_S):
-        watts = peak * math.sin(math.pi * t / duration)
-        events.append(Event(at=float(t), apply=_set_dc_input(battery_index, watts)))
-    return events
-
-
-def _household_and_solar(
-    rng: random.Random, duration: float, solar_peak: float
-) -> list[Event]:
-    """Household appliance steps over an AC solar day curve with cloud dips.
-
-    Combines the discharge-side step schedule with a half-sine PV curve big
-    enough to push the pool into export/charge territory around midday, so the
-    scenario exercises the full bidirectional loop (charge distribution, the
-    AC-charge clamp, zero-crossings) on top of the step responses — not just
-    discharge.
-    """
-    return (
-        _household_steps(rng, duration)
-        + _solar_curve(duration, solar_peak)
-        + _cloud_dips(rng, duration)
-    )
-
-
-def build_scenarios() -> dict[str, Scenario]:
-    """All evaluation scenarios, keyed by name.
-
-    Multi-battery scenarios come in two balancer modes: plain fair-share
-    (``…/fair``) and efficiency optimization (``…/eff``, exercising
-    deprioritization, rotation, saturation swaps and probe handoffs).
-    """
-    scenarios: dict[str, Scenario] = {}
-
-    def add(s: Scenario) -> None:
-        scenarios[s.name] = s
-
-    dur_steps = 3600.0
-    add(
-        Scenario(
-            name="single_venus_steps",
-            description="One Venus, stepped house load (kettle/oven/dishwasher)",
-            batteries=[_VENUS],
-            duration_s=dur_steps,
-            loads=list(_HOUSEHOLD_LOADS),
-            build_events=lambda rng: _household_steps(rng, dur_steps),
-        )
-    )
-
-    add(
-        Scenario(
-            name="single_venus_steps_slow",
-            description=(
-                "One Venus, stepped house load over a slow meter (fresh reading "
-                "only every 10 s + ~1 s delay) — coarse-sampling stress, cf. "
-                "single_venus_steps"
-            ),
-            batteries=[_VENUS],
-            duration_s=dur_steps,
-            loads=list(_HOUSEHOLD_LOADS),
-            build_events=lambda rng: _household_steps(rng, dur_steps),
-            meter_interval_s=10.0,
-            meter_latency_s=1.0,
-        )
-    )
-
-    dur_washer = 1800.0
-    add(
-        Scenario(
-            name="single_venus_washer",
-            description=(
-                "One Venus, washing-machine drum tumble (~120 W motor "
-                "pausing/restarting every ~16 s) over a meter with ~1 s "
-                "latency — sustained-oscillation stress (issue #473)"
-            ),
-            batteries=[_VENUS],
-            duration_s=dur_washer,
-            base_load=[450.0, 0.0, 0.0],
-            loads=list(_WASHER_LOADS),
-            build_events=lambda rng: _washer_cycle(rng, dur_washer),
-            # The field setup read an HA push sensor with measurement+transport
-            # delay; that latency is what turns each drum disturbance into a
-            # loop that hunts continuously instead of settling between pauses
-            # (issue #473). Without it the loop settles into ~10 s calm windows
-            # the real trace never showed.
-            meter_latency_s=1.0,
-        )
-    )
-
-    add(
-        Scenario(
-            name="single_venus_noisy",
-            description=(
-                "One Venus, pretty noisy house baseline "
-                f"(±{_NOISY_BASE_NOISE:g} W jitter, no discrete appliance steps) "
-                "— noise-rejection stress"
-            ),
-            batteries=[_VENUS],
-            duration_s=dur_steps,
-            base_load=list(_NOISY_BASE_LOAD),
-            base_noise=_NOISY_BASE_NOISE,
-            build_events=_no_events,
-        )
-    )
-
-    add(
-        Scenario(
-            name="single_venus_trace",
-            description=(
-                "One Venus, real recorded household load (RAE House 1, 1 Hz "
-                "trace, CC BY) over a meter with realistic ~0.8 s latency — "
-                "real-world correlated-load stress (cf. the synthetic-noise "
-                "single_venus_noisy)"
-            ),
-            batteries=[_VENUS],
-            duration_s=dur_steps,
-            base_load=[_household_trace()[0][1], 0.0, 0.0],
-            base_noise=_TRACE_METER_NOISE,
-            build_events=lambda rng: _trace_events(rng, dur_steps),
-            # Real installs read the meter with measurement+transport delay;
-            # pairing the real load with realistic latency is the field condition
-            # the synthetic latency-free scenarios never cover.
-            meter_latency_s=0.8,
-        )
-    )
-
-    dur_solar = 5400.0
-    solar_peak = 1800.0
-    add(
-        Scenario(
-            name="single_venus_solar",
-            description="One Venus, solar day curve crossing into export + clouds",
-            batteries=[BatterySpec(initial_soc=0.4)],
-            duration_s=dur_solar,
-            base_load=[400.0, 0.0, 0.0],
-            build_events=lambda rng: (
-                _solar_curve(dur_solar, solar_peak) + _cloud_dips(rng, dur_solar)
-            ),
-        )
-    )
-
-    add(
-        Scenario(
-            name="single_venus_solar_slow",
-            description=(
-                "One Venus, solar day + clouds over a slow meter (fresh reading "
-                "only every 10 s + ~1 s delay) — coarse sampling of a moving PV "
-                "setpoint, cf. single_venus_solar"
-            ),
-            batteries=[BatterySpec(initial_soc=0.4)],
-            duration_s=dur_solar,
-            base_load=[400.0, 0.0, 0.0],
-            build_events=lambda rng: (
-                _solar_curve(dur_solar, solar_peak) + _cloud_dips(rng, dur_solar)
-            ),
-            meter_interval_s=10.0,
-            meter_latency_s=1.0,
-        )
-    )
-
-    # SoC-saturation scenarios: short runs against a 5 kWh pack barely move the
-    # SoC, so the empty/full handoff (battery stops discharging/charging and the
-    # grid must take over) was never exercised. These use a realistic single-unit
-    # 2.56 kWh pack started near an edge and driven hard enough to hit it.
-    dur_drain = 10800.0  # 3 h
-    add(
-        Scenario(
-            name="single_venus_drain",
-            description=(
-                "One Venus emptying under a sustained real evening load (2.56 kWh "
-                "pack, low initial SoC) — empty-saturation handoff to grid import"
-            ),
-            batteries=[BatterySpec(capacity_wh=2560.0, initial_soc=0.35)],
-            duration_s=dur_drain,
-            base_load=[_household_trace()[0][1], 0.0, 0.0],
-            base_noise=_TRACE_METER_NOISE,
-            build_events=lambda rng: _trace_events(rng, dur_drain),
-            meter_latency_s=0.5,
-        )
-    )
-
-    add(
-        Scenario(
-            name="single_venus_fill",
-            description=(
-                "One Venus filling under a strong solar day (2.56 kWh pack, high "
-                "initial SoC) — full-saturation handoff: charge clamp + export"
-            ),
-            batteries=[BatterySpec(capacity_wh=2560.0, initial_soc=0.8)],
-            duration_s=dur_solar,
-            base_load=[300.0, 0.0, 0.0],
-            build_events=lambda rng: (
-                _solar_curve(dur_solar, 2200.0) + _cloud_dips(rng, dur_solar)
-            ),
-            meter_latency_s=0.5,
-        )
-    )
-
-    # Real PV net-load: matching recorded PV + load from one Cyprus prosumer on a
-    # partly-cloudy day, replacing the synthetic half-sine + scripted cloud dips
-    # with genuine PV cloud transients. Exercises the bidirectional loop (charge
-    # on export, discharge on cloud dips) against real solar variability.
-    add(
-        Scenario(
-            name="single_venus_pv",
-            description=(
-                "One Venus, real recorded PV + load net-load (Cyprus prosumer, "
-                "30 s, real cloud transients) — bidirectional charge/export "
-                "tracking, cf. the synthetic single_venus_solar"
-            ),
-            batteries=[BatterySpec(initial_soc=0.4)],
-            duration_s=dur_solar,
-            base_load=[_net_trace()[0][1] * _PV_SCALE, 0.0, 0.0],
-            base_noise=_TRACE_METER_NOISE,
-            build_events=lambda rng: _pv_net_events(rng, dur_solar),
-            meter_latency_s=0.5,
-        )
-    )
-
-    add(
-        Scenario(
-            name="single_venus_d_steps",
-            description="One Venus D (VNSD-0 integer loop), stepped house load",
-            batteries=[_VENUS_D],
-            duration_s=dur_steps,
-            loads=list(_HOUSEHOLD_LOADS),
-            build_events=lambda rng: _household_steps(rng, dur_steps),
-        )
-    )
-
-    add(
-        Scenario(
-            name="single_venus_d_washer",
-            description=(
-                "One Venus D, washing-machine drum tumble over a ~1 s-latency "
-                "meter — sustained-oscillation stress for the integer loop "
-                "(±11 W deadband, no spike filter), cf. single_venus_washer"
-            ),
-            batteries=[_VENUS_D],
-            duration_s=dur_washer,
-            base_load=[450.0, 0.0, 0.0],
-            loads=list(_WASHER_LOADS),
-            build_events=lambda rng: _washer_cycle(rng, dur_washer),
-            meter_latency_s=1.0,
-        )
-    )
-
-    add(
-        Scenario(
-            name="single_venus_d_solar",
-            description="One Venus D, solar day curve crossing into export + clouds",
-            batteries=[BatterySpec(device_type="VNSD-0", initial_soc=0.4)],
-            duration_s=dur_solar,
-            base_load=[400.0, 0.0, 0.0],
-            build_events=lambda rng: (
-                _solar_curve(dur_solar, solar_peak) + _cloud_dips(rng, dur_solar)
-            ),
-        )
-    )
-
-    # Heterogeneous phase: a Venus D (integer integrator) sharing one phase with
-    # a Venus C (HMG float ramp), so the two different control laws coexist
-    # under the same balancer.
-    for mode, kwargs in (("fair", {}), ("eff", _EFF_MODE)):
-        add(
-            Scenario(
-                name=f"venus_d_plus_c/{mode}",
-                description="One Venus D + one Venus C sharing one phase",
-                batteries=[_VENUS_D, _VENUS],
-                duration_s=dur_steps,
-                loads=list(_HOUSEHOLD_LOADS),
-                build_events=lambda rng: _household_steps(rng, dur_steps),
-                ct_kwargs=dict(kwargs),
-            )
-        )
-
-    for mode, kwargs in (("fair", {}), ("eff", _EFF_MODE)):
-        add(
-            Scenario(
-                name=f"two_venus/{mode}",
-                description="Two identical Venus sharing one phase",
-                batteries=[_VENUS, _VENUS],
-                duration_s=dur_steps,
-                loads=list(_HOUSEHOLD_LOADS),
-                build_events=lambda rng: _household_steps(rng, dur_steps),
-                ct_kwargs=dict(kwargs),
-            )
-        )
-
-    add(
-        Scenario(
-            name="two_venus_slow/fair",
-            description=(
-                "Two Venus sharing one phase, stepped load over a slow meter "
-                "(fresh reading only every 10 s + ~1 s delay) — coarse sampling "
-                "with share-splitting, cf. two_venus/fair"
-            ),
-            batteries=[_VENUS, _VENUS],
-            duration_s=dur_steps,
-            loads=list(_HOUSEHOLD_LOADS),
-            build_events=lambda rng: _household_steps(rng, dur_steps),
-            meter_interval_s=10.0,
-            meter_latency_s=1.0,
-        )
-    )
-
-    # Three-phase imbalance: one Venus per phase with asymmetric per-phase base
-    # loads and a different appliance on each phase. Everything else in the suite
-    # is single-phase (load + batteries on A); this exercises the active-control
-    # loop's per-phase target distribution — each unit must null only its own
-    # phase, with no cross-phase interference.
-    add(
-        Scenario(
-            name="phase_imbalance",
-            description=(
-                "Three Venus, one per phase; asymmetric per-phase base + a "
-                "different appliance on each phase — per-phase distribution"
-            ),
-            batteries=[
-                BatterySpec(phase="A"),
-                BatterySpec(phase="B"),
-                BatterySpec(phase="C"),
-            ],
-            duration_s=dur_steps,
-            base_load=[300.0, 200.0, 150.0],
-            loads=list(_PHASE_IMBALANCE_LOADS),
-            build_events=lambda rng: _household_steps(rng, dur_steps),
-        )
-    )
-
-    for mode, kwargs in (("fair", {}), ("eff", _EFF_MODE)):
-        add(
-            Scenario(
-                name=f"two_venus_noisy/{mode}",
-                description=(
-                    "Two Venus sharing one phase, pretty noisy house baseline "
-                    f"(±{_NOISY_BASE_NOISE:g} W jitter, no discrete appliance steps)"
-                ),
-                batteries=[_VENUS, _VENUS],
-                duration_s=dur_steps,
-                base_load=list(_NOISY_BASE_LOAD),
-                base_noise=_NOISY_BASE_NOISE,
-                build_events=_no_events,
-                ct_kwargs=dict(kwargs),
-            )
-        )
-
-    # fair always splits the load across both units; eff (with the raised
-    # _TRACE_EFF_MODE floor) concentrates on one Venus during the calm base and
-    # lets the second cut in on peaks — so the two variants differ under a real
-    # load instead of /eff duplicating /fair (which it does at the default floor).
-    for mode, kwargs in (("fair", {}), ("eff", _TRACE_EFF_MODE)):
-        suffix = (
-            "fair-share splitting across both units"
-            if mode == "fair"
-            else "efficiency concentration: the 2nd Venus cuts in only on peaks"
-        )
-        add(
-            Scenario(
-                name=f"two_venus_trace/{mode}",
-                description=(
-                    "Two Venus sharing one phase, real recorded household load "
-                    "(RAE House 1, 1 Hz trace, CC BY) over a ~0.8 s-latency "
-                    f"meter — {suffix}"
-                ),
-                batteries=[_VENUS, _VENUS],
-                duration_s=dur_steps,
-                base_load=[_household_trace()[0][1], 0.0, 0.0],
-                base_noise=_TRACE_METER_NOISE,
-                build_events=lambda rng: _trace_events(rng, dur_steps),
-                meter_latency_s=0.8,
-                ct_kwargs=dict(kwargs),
-            )
-        )
-
-    # Solar peak (W) for the multi-Venus solar scenarios: above the base load
-    # plus typical appliance draw, so midday PV pushes the pool into charging /
-    # export for stretches.
-    solar_peak_house = 3000.0
-    for mode, kwargs in (("fair", {}), ("eff", _EFF_MODE)):
-        add(
-            Scenario(
-                name=f"two_venus_solar/{mode}",
-                description="Two Venus, household load + solar day curve + clouds",
-                batteries=[_VENUS, _VENUS],
-                duration_s=dur_solar,
-                base_load=[400.0, 0.0, 0.0],
-                loads=list(_HOUSEHOLD_LOADS),
-                build_events=lambda rng: _household_and_solar(
-                    rng, dur_solar, solar_peak_house
-                ),
-                ct_kwargs=dict(kwargs),
-            )
-        )
-
-    # Issue #522: fair distribution, both units active, one FULL (cannot charge)
-    # on the PV-export phase while the healthy unit (on another phase) is already
-    # charging. Under a sustained surplus the balance correction pulls the
-    # healthy unit's command down toward an even split with the idle full unit,
-    # so part of its absorption falls inside the battery firmware's input
-    # deadband and is held — the healthy unit under-charges and the surplus is
-    # partly exported. The fix lets the full unit be recognised as saturated
-    # (dropping out of the share math) so the healthy unit keeps more of the
-    # command above the deadband and absorbs more. Ramp pacing's base step is
-    # below the saturation min-target (the reporter's pace_base_step <
-    # min_target_for_saturation), the condition under which the full unit's
-    # pinned command used to read as idle and never registered as saturated.
-    # The effect requires a unit already in motion (initial_power): the firmware
-    # deadband holds an in-motion sub-deadband command, where a cold start would
-    # instead just sit at rest, so without it the handoff dynamics differ and the
-    # fix shows no benefit.
-    add(
-        Scenario(
-            name="full_battery_low_pace",
-            description=(
-                "Fair distribution, both active: a full unit on the PV-export "
-                "phase + a healthy unit (already charging) on another phase, "
-                "sustained surplus; ramp-pace base step below the saturation "
-                "min-target (issue #522). The full unit must be detected "
-                "saturated so the healthy one keeps absorbing instead of having "
-                "part of its command held inside the firmware deadband"
-            ),
-            batteries=[
-                BatterySpec(
-                    device_type="VNSE3-0",
-                    phase="A",
-                    max_charge_power=0,  # full: cannot charge
-                    initial_soc=0.95,
-                ),
-                BatterySpec(
-                    device_type="VNSE3-0",
-                    phase="B",
-                    initial_soc=0.4,
-                    initial_power=-200.0,  # already charging when the run begins
-                ),
-            ],
-            duration_s=900.0,
-            base_load=[-250.0, 0.0, 0.0],  # ~250 W sustained PV surplus
-            build_events=lambda rng: [],
-            ct_kwargs={
-                "balance_gain": 0.40,
-                "balance_deadband": 30.0,
-                "max_correction_per_step": 150.0,
-                "pace_base_step": 30.0,
-                "pace_max_step": 200.0,
-                "min_target_for_saturation": 40.0,
-                "saturation_alpha": 0.9,
-            },
-            meter_latency_s=0.5,
-        )
-    )
-
-    # Two B2500 on one phase with MIN_DC_OUTPUT set *below* the ~80 W a
-    # channel pair needs to start -- the configuration issue #600 reports.
-    # The floor is where the balancer parks a starved unit, so a value under
-    # the start floor parks it at a command it cannot execute; whether that
-    # counts as evidence of saturation decides if the second unit ever gets a
-    # real share or sits at 0 W while the first carries the house.  Nothing
-    # else in the suite sets MIN_DC_OUTPUT, so this is the only scenario that
-    # exercises the floor at all.
-    dur_floor = 3600.0
-    add(
-        Scenario(
-            name="b2500_pair_dc_floor",
-            description="Two DC-only B2500 on one phase, MIN_DC_OUTPUT below "
-            "their start floor",
-            batteries=[
-                # One unit already carrying the house, the other at rest --
-                # the state a restart leaves behind, and the one the reporter
-                # captured (230 W against 4 W).  Starting both from zero would
-                # split the first correction evenly and neither would ever be
-                # starved.
-                replace(_B2500_PAIR, initial_power=250.0),
-                _B2500_PAIR,
-            ],
-            duration_s=dur_floor,
-            base_load=[250.0, 0.0, 0.0],
-            loads=list(_HOUSEHOLD_LOADS),
-            build_events=lambda rng: _household_steps(rng, dur_floor),
-            ct_kwargs={"min_dc_output": 20.0},
-        )
-    )
-
-    dur_mixed = 5400.0
-    for mode, kwargs in (("fair", {}), ("eff", _EFF_MODE)):
-        add(
-            Scenario(
-                name=f"mixed_venus_b2500/{mode}",
-                description="Two Venus + one DC-only B2500 with PV input",
-                batteries=[_VENUS, _VENUS, _B2500],
-                duration_s=dur_mixed,
-                loads=list(_HOUSEHOLD_LOADS),
-                build_events=lambda rng: (
-                    _household_steps(rng, dur_mixed)
-                    + _dc_solar_curve(dur_mixed, 700.0, battery_index=2)
-                ),
-                ct_kwargs=dict(kwargs),
-            )
-        )
-
-    for mode, kwargs in (("fair", {}), ("eff", _EFF_MODE)):
-        add(
-            Scenario(
-                name=f"mixed_cadence/{mode}",
-                description="Slow-polling V2 (3.1 s) + fast V3 (0.45 s)",
-                batteries=[_VENUS_V2_SLOW, _VENUS_V3],
-                duration_s=dur_steps,
-                loads=list(_HOUSEHOLD_LOADS),
-                build_events=lambda rng: _household_steps(rng, dur_steps),
-                ct_kwargs=dict(kwargs),
-            )
-        )
-
-    for mode, kwargs in (("fair", {}), ("eff", _EFF_MODE)):
-        add(
-            Scenario(
-                name=f"mixed_cadence_solar/{mode}",
-                description="Slow V2 + fast V3, household load + solar + clouds",
-                batteries=[_VENUS_V2_SLOW, _VENUS_V3],
-                duration_s=dur_solar,
-                base_load=[400.0, 0.0, 0.0],
-                loads=list(_HOUSEHOLD_LOADS),
-                build_events=lambda rng: _household_and_solar(
-                    rng, dur_solar, solar_peak_house
-                ),
-                ct_kwargs=dict(kwargs),
-            )
-        )
-
-    return scenarios
-
-
-# ---------------------------------------------------------------------------
-# Reporting
-# ---------------------------------------------------------------------------
-
-# Metrics shown in tables, in order, with "lower is better" direction (all
-# current metrics improve downward).
-_REPORT_METRICS = [
-    "settle_mean_s",
-    "settle_p95_s",
-    "unsettled_events",
-    "overshoot_mean_w",
-    "overshoot_max_w",
-    "band_crossings_per_h",
-    "grid_p2p_w",
-    "grid_rms_w",
-    "steady_rms_w",
-    "mean_abs_grid_w",
-    "share_imbalance_w",
-    "avoidable_import_wh",
-    "avoidable_export_wh",
-    "cost_regret_ct",
-    "battery_travel_w_per_h",
-]
-
-# Real-world priority weights for the *priority-weighted* aggregate score, on
-# top of the equal-weighted roll-up.  The flat mean treats a 1% move in
-# worst-case overshoot the same as a 1% move in 95th-percentile settle time;
-# in the field those are worth very different amounts.  These weights encode the
-# value hierarchy a self-consumption controller is actually judged on:
-#
-#   * **Cost regret vs perfect foresight is the money north-star** — the single
-#     ungameable scalar (real bill minus the optimal-dispatch bill), so it
-#     carries top weight alongside import.
-#   * **Self-consumption energy is the product's purpose**, and the two sides
-#     are not symmetric: avoidable grid *import* is paid at the retail tariff
-#     while avoidable *export* earns only the (much lower) feed-in tariff, so
-#     import is weighted ~4x export.  (``cost_regret_ct`` is the fused money view
-#     of the same effect; the two energy metrics keep their finer-grained lens.)
-#   * **Sign-flip overshoot and sustained hunting are do-no-harm guardrails** —
-#     overshoot flips the grid sign (actively worse than no battery) and hunting
-#     wastes energy, wears the pack and reads as "broken" — so they carry heavy
-#     weight (and a separate hard regression flag, see ``_GUARDRAIL_METRICS``).
-#   * **Battery travel is cycle-life / hardware wear**, a real € cost.
-#   * **Inter-battery share imbalance is cycle-life / fairness** — a lopsided
-#     split (issue #523) wears and saturates one pack ahead of its peers, so it
-#     shares the cycle-life weight with battery travel.
-#   * **Settle time is an enabler** (it mostly feeds energy), weighted lightly.
-#
-# All metrics are lower-is-better, so a negative weighted change is an overall
-# improvement.  Weights are relative (only their ratios matter).
-_METRIC_WEIGHTS: dict[str, float] = {
-    "cost_regret_ct": 4.0,
-    "avoidable_import_wh": 4.0,
-    "avoidable_export_wh": 1.0,
-    "overshoot_max_w": 3.0,
-    "band_crossings_per_h": 2.0,
-    "battery_travel_w_per_h": 2.0,
-    "share_imbalance_w": 2.0,
-    "grid_p2p_w": 1.5,
-    "grid_rms_w": 1.0,
-    "overshoot_mean_w": 1.0,
-    "mean_abs_grid_w": 1.0,
-    "steady_rms_w": 0.5,
-    "settle_mean_s": 0.5,
-    "unsettled_events": 0.5,
-    "settle_p95_s": 0.3,
-}
-
-# Do-no-harm guardrails: regressing one of these makes the system *actively
-# worse than doing nothing*.  Overshoot flips the grid sign; band crossings /
-# peak-to-peak are sustained hunting; avoidable import is missed self-consumption
-# (energy imported while a battery could have supplied it) — the product's
-# purpose, paid at the retail tariff and free to fix (just discharge).  Cost
-# regret is the headline money metric — the bill the controller ran up over what
-# perfect foresight would have paid — ungameable (both grid directions cost) and
-# 0 only when the controller matched the optimum, so a regression there is real
-# money lost regardless of which component moved.  A regression beyond
-# ``_GUARDRAIL_TOLERANCE`` (5%) is surfaced as a hard warning regardless of how
-# good the weighted score looks, so a smoother controller can't silently trade
-# self-consumption for stability.
-#
-# Avoidable *export* is deliberately NOT a guardrail (it keeps its 1.0 soft
-# weight in ``_METRIC_WEIGHTS``).  The metric conflates two behaviours: free-to-
-# fix over-discharge (a battery discharging *into* a grid export because it
-# wasn't throttled when load dropped — genuine do-no-harm) and a legitimate
-# missed-AC-charging trade (letting a surplus export rather than pay a charge
-# round-trip).  A hard flag on the conflated sum would penalise the second,
-# legitimate case; it is also gated on ``ac_chargeable`` so it misses over-
-# discharge entirely for DC-only packs (e.g. B2500), and it dominates / gets
-# noisy in steady solar-surplus scenarios.  A future device-agnostic
-# over-discharge-only metric could be guardrailed cleanly.
-#
-# ``avoidable_import_wh`` and ``cost_regret_ct`` also legitimately sit at 0 in
-# the base (a controller already matching perfect foresight), so a base of 0
-# going positive is itself the regression — see ``_guardrail_regressions`` for
-# the zero-base handling.
-_GUARDRAIL_METRICS = (
-    "overshoot_max_w",
-    "band_crossings_per_h",
-    "grid_p2p_w",
-    "avoidable_import_wh",
-    "cost_regret_ct",
-)
-_GUARDRAIL_TOLERANCE = 0.05
-
-# Short, human-readable description for each metric in `_REPORT_METRICS`,
-# rendered as a collapsible glossary in the CI PR comment. Keep in sync with
-# `_REPORT_METRICS` and the metric computation in `_score()`.
-_METRIC_GLOSSARY = [
-    (
-        "settle_mean_s",
-        f"Mean seconds after a load/PV step for grid power to return inside the "
-        f"±{SETTLE_BAND_W:g} W settle band and hold for {SETTLE_HOLD_S:g} s "
-        f"(reaction speed).",
-    ),
-    (
-        "settle_p95_s",
-        "95th-percentile settle time — the slow tail of reactions.",
-    ),
-    (
-        "unsettled_events",
-        f"Number of disturbance events that never settled within the "
-        f"{EVENT_WINDOW_S / 60:g}-minute measurement window.",
-    ),
-    (
-        "overshoot_mean_w",
-        "Mean overshoot (W): how far grid power swings past zero to the "
-        "opposite sign after an event.",
-    ),
-    (
-        "overshoot_max_w",
-        "Worst-case overshoot (W) across all events.",
-    ),
-    (
-        "band_crossings_per_h",
-        f"Sign flips per hour across the ±{OSC_BAND_W:g} W hysteresis band — "
-        f"oscillation / hunting frequency.",
-    ),
-    (
-        "grid_p2p_w",
-        "Sustained peak-to-peak grid swing (95th - 5th percentile) over the "
-        "whole run — oscillation amplitude. Non-zero whenever the loop keeps "
-        "hunting, including continuous oscillation the step-response metrics "
-        "(settle/overshoot) miss.",
-    ),
-    (
-        "grid_rms_w",
-        "RMS grid power (W) over the *whole* run, transients included — the L2 "
-        "tracking error: how cleanly the loop held zero, penalising big "
-        "excursions (overshoot, swings) far harder than a small steady offset. "
-        "Pairs with battery_travel_w_per_h as the control-effort term.",
-    ),
-    (
-        "steady_rms_w",
-        f"RMS grid power (W) during steady state (excluding the "
-        f"{STEADY_EXCLUDE_S:g} s after each event) — residual jitter when "
-        f"nothing is changing.",
-    ),
-    (
-        "mean_abs_grid_w",
-        "Mean absolute grid power (W) over the whole run — overall tracking accuracy.",
-    ),
-    (
-        "share_imbalance_w",
-        "Time-weighted watts misallocated between batteries sharing a phase (sum "
-        "of each battery's deviation from the even fair share) — 0 when the pool "
-        "splits load evenly, higher when one battery is left lopsided (issue "
-        "#523). 0 for scenarios with at most one battery per phase.",
-    ),
-    (
-        "avoidable_import_wh",
-        "Energy imported from the grid (Wh) the battery could have supplied "
-        "(it had charge and discharge headroom) — missed self-consumption.",
-    ),
-    (
-        "avoidable_export_wh",
-        "Energy exported to the grid (Wh) an AC-chargeable battery could have "
-        "absorbed (it had room and charge headroom) — missed charging.",
-    ),
-    (
-        "cost_regret_ct",
-        f"Money north-star: electricity bill (eurocents, import @ "
-        f"{RETAIL_CT_PER_KWH:g} ct/kWh, export @ {FEEDIN_CT_PER_KWH:g} ct/kWh) "
-        f"over what a perfect-foresight optimal battery would have paid on the "
-        f"same load. Ungameable (both grid directions cost); 0 = matched the "
-        f"optimum. The single number that says how much the controller left on "
-        f"the table.",
-    ),
-    (
-        "battery_travel_w_per_h",
-        "Total absolute change in battery setpoints per hour (W/h) — control "
-        "effort / actuator wear; lower is smoother.",
-    ),
-]
-
-
-def _metric_ndp(key: str) -> int:
-    """Decimal places to round a metric to. Eurocent costs need 2: per-scenario
-    ``cost_regret_ct`` clusters at fractions of a cent, so rounding the
-    aggregate/seed-merge to 1 dp would quantize the guardrail's 5% test into
-    pure noise. Every other metric keeps the 1-dp report convention."""
-    return 2 if key.endswith("_ct") else 1
-
-
-def _mean_value(values: list, ndp: int = 1):
-    """Average a homogeneous list of result values across seeds.
-
-    Recurses into nested lists (a battery's trace, the list of per-battery
-    traces) so traces average element-by-element; string lists (battery
-    labels, identical across seeds) pass through as the first value. *ndp* is
-    the rounding precision for numeric values (see :func:`_metric_ndp`)."""
-    v0 = values[0]
-    if isinstance(v0, bool):
-        return v0
-    if isinstance(v0, (int, float)):
-        return round(sum(float(v) for v in values) / len(values), ndp)
-    if isinstance(v0, list):
-        if v0 and isinstance(v0[0], str):
-            return v0  # labels are identical across seeds
-        return [_mean_value([v[i] for v in values], ndp) for i in range(len(v0))]
-    return v0  # strings / anything else: identical across seeds
-
-
-def _merge_seeds(per_seed: list[dict]) -> dict:
-    """Collapse one scenario's per-seed results into a single averaged row.
-
-    Every numeric metric (and every trace, element-wise) becomes the mean over
-    the seeds, so the reported figure is the seed-averaged signal rather than
-    one noisy draw. A lone seed is returned unchanged (no averaging, keeps the
-    original ``seed`` field). The merged row carries ``seeds`` / ``n_seeds``
-    instead of a single ``seed``."""
-    if len(per_seed) == 1:
-        return per_seed[0]
-    merged: dict = {
-        "scenario": per_seed[0]["scenario"],
-        "seeds": [r["seed"] for r in per_seed],
-        "n_seeds": len(per_seed),
-    }
-    for key in per_seed[0]:
-        if key in ("scenario", "seed"):
-            continue
-        merged[key] = _mean_value([r[key] for r in per_seed], _metric_ndp(key))
-    return merged
-
-
-def _aggregate(results: list[dict]) -> dict:
-    """Collapse a result list into one synthetic "AGGREGATE" row.
-
-    Each reported metric becomes its **unweighted mean across the scenarios**
-    that carry it (a base produced before a metric existed simply doesn't
-    contribute that key — it's omitted from the aggregate, mirroring the
-    per-scenario renderers' ``—`` handling). One row that answers "did this
-    change help *overall*?" at a glance, on top of the per-scenario tables.
-
-    Means are scale-mixing on purpose: they're a rough headline, while the
-    per-metric relative deltas (and :func:`_overall_summary`) are the
-    unit-independent read on direction.
-    """
-    agg: dict = {"scenario": "AGGREGATE", "n_scenarios": len(results)}
-    for key in _REPORT_METRICS:
-        vals = [float(r[key]) for r in results if key in r]
-        if vals:
-            agg[key] = round(sum(vals) / len(vals), _metric_ndp(key))
-    return agg
-
-
-def _compare_aggregates(
-    base: list[dict] | None, head: list[dict]
-) -> tuple[dict | None, dict]:
-    """Aggregate base and head for a base-vs-head comparison.
-
-    With a baseline, both sides are aggregated over the **scenarios they share**
-    so the Overall verdict and aggregate table compare like for like even when
-    the two runs cover different scenario sets (e.g. a PR adds or drops a
-    scenario). Without a baseline, head is aggregated over all its scenarios."""
-    if not base:
-        return None, _aggregate(head)
-    shared = {r["scenario"] for r in base} & {r["scenario"] for r in head}
-    base_agg = _aggregate([r for r in base if r["scenario"] in shared])
-    head_agg = _aggregate([r for r in head if r["scenario"] in shared])
-    return base_agg, head_agg
-
-
-def _overall_change(
-    base_agg: dict, head_agg: dict
-) -> tuple[int, int, int, float | None]:
-    """Count improved / regressed / unchanged aggregate metrics and the mean
-    relative change across them (``None`` when nothing is comparable).
-
-    Every reported metric is lower-is-better, so a negative mean is an overall
-    improvement. The mean is over *relative* deltas (head vs base, per metric),
-    making it unit-independent; metrics whose base aggregate is 0 contribute to
-    the up/down counts but not the percentage (no defined relative change)."""
-    improved = regressed = unchanged = 0
-    deltas: list[float] = []
-    for key in _REPORT_METRICS:
-        if key not in base_agg or key not in head_agg:
-            continue
-        bv, hv = float(base_agg[key]), float(head_agg[key])
-        if hv < bv:
-            improved += 1
-        elif hv > bv:
-            regressed += 1
-        else:
-            unchanged += 1
-        if bv != 0:
-            deltas.append((hv - bv) / abs(bv))
-    mean_pct = sum(deltas) / len(deltas) * 100.0 if deltas else None
-    return improved, regressed, unchanged, mean_pct
-
-
-def _overall_summary(base_agg: dict, head_agg: dict) -> str:
-    """One-line verdict for the aggregate: how many metrics moved each way and
-    the mean relative change (lower is better)."""
-    improved, regressed, unchanged, mean_pct = _overall_change(base_agg, head_agg)
-    if mean_pct is None:
-        trend = "no comparable metrics"
-    elif mean_pct < 0:
-        trend = f"mean {mean_pct:.1f}% (better)"
-    elif mean_pct > 0:
-        trend = f"mean +{mean_pct:.1f}% (worse)"
-    else:
-        trend = "mean 0% (unchanged)"
-    return (
-        f"{improved} improved, {regressed} regressed, {unchanged} unchanged "
-        f"across {improved + regressed + unchanged} metrics — {trend}"
-    )
-
-
-def _weighted_overall(base_agg: dict, head_agg: dict) -> float | None:
-    """Priority-weighted mean relative change across aggregate metrics.
-
-    Like :func:`_overall_change`'s percentage but weighted by
-    :data:`_METRIC_WEIGHTS` so the headline reflects real-world value (import-
-    heavy energy, guardrail overshoot/hunting, cycle-life travel) rather than
-    treating every metric equally.  Lower is better, so a negative result is an
-    improvement; ``None`` when nothing comparable has a defined relative change.
-    """
-    num = den = 0.0
-    for key, weight in _METRIC_WEIGHTS.items():
-        if key not in base_agg or key not in head_agg:
-            continue
-        bv, hv = float(base_agg[key]), float(head_agg[key])
-        if bv == 0:
-            continue
-        num += weight * (hv - bv) / abs(bv)
-        den += weight
-    if den == 0:
-        return None
-    return num / den * 100.0
-
-
-def _guardrail_regressions(base_agg: dict, head_agg: dict) -> list[str]:
-    """Do-no-harm guardrail metrics (:data:`_GUARDRAIL_METRICS`) that regressed
-    beyond :data:`_GUARDRAIL_TOLERANCE`, formatted as ``metric +N%`` (or
-    ``metric 0→N`` when the base was zero).
-
-    These are surfaced as a hard warning independent of the weighted score: a
-    controller that wins on average but flips the grid sign harder, hunts more,
-    or imports more from the grid has made the system worse where it matters
-    most.
-
-    A zero base needs explicit handling: a relative change against 0 is
-    undefined, so the old ``bv > 0`` filter silently dropped the worst possible
-    breach — a metric going from *nothing* to *something* (e.g. zero overshoot
-    becoming a real sign flip, or a perfectly self-consuming controller starting
-    to import). That transition is now flagged outright as ``0→N``."""
-    out: list[str] = []
-    for key in _GUARDRAIL_METRICS:
-        if key not in base_agg or key not in head_agg:
-            continue
-        bv, hv = float(base_agg[key]), float(head_agg[key])
-        if bv == 0:
-            if hv > 0:
-                out.append(f"{key} 0→{hv:g}")
-        elif (hv - bv) / bv > _GUARDRAIL_TOLERANCE:
-            out.append(f"{key} +{(hv - bv) / bv * 100.0:.0f}%")
-    return out
-
-
-def _priority_summary(base_agg: dict, head_agg: dict) -> str:
-    """One-line priority-weighted verdict plus the do-no-harm guardrail status.
-
-    Complements :func:`_overall_summary` (the equal-weighted roll-up) with the
-    weighted score and an explicit guardrail flag — see :data:`_METRIC_WEIGHTS`
-    and :data:`_GUARDRAIL_METRICS`."""
-    pct = _weighted_overall(base_agg, head_agg)
-    if pct is None:
-        score = "no comparable metrics"
-    elif pct < 0:
-        score = f"{pct:.1f}% (better)"
-    elif pct > 0:
-        score = f"+{pct:.1f}% (worse)"
-    else:
-        score = "0% (unchanged)"
-    regressions = _guardrail_regressions(base_agg, head_agg)
-    if regressions:
-        guard = "⚠️ do-no-harm guardrail regressed: " + ", ".join(regressions)
-    else:
-        guard = "✅ no do-no-harm guardrail regressions"
-    return f"priority-weighted {score} — {guard}"
-
-
-def render_text(results: list[dict]) -> str:
-    lines = []
-    # A single aggregate row up top makes an overall read possible without
-    # eyeballing every scenario (skipped for a lone scenario — it'd just echo
-    # that scenario's own numbers).
-    if len(results) > 1:
-        agg = _aggregate(results)
-        lines.append(f"== AGGREGATE (mean across {agg['n_scenarios']} scenarios)")
-        for key in _REPORT_METRICS:
-            if key in agg:
-                lines.append(f"  {key:<24} {agg[key]}")
-        lines.append("")
-    for res in results:
-        lines.append(
-            f"== {res['scenario']} ({_seed_label(res)}, "
-            f"{res['duration_h']}h, {res['events_measured']} events)"
-        )
-        for key in _REPORT_METRICS:
-            lines.append(f"  {key:<24} {res[key]}")
-    return "\n".join(lines)
-
-
-def _seed_label(res: dict) -> str:
-    """How a result's seed provenance reads in a header: a single ``seed N`` or
-    ``mean of N seeds`` once per-seed results have been merged."""
-    if res.get("n_seeds"):
-        return f"mean of {res['n_seeds']} seeds"
-    return f"seed {res.get('seed', '?')}"
-
-
-def _fmt_delta(base: float, head: float) -> str:
-    if base == head:
-        return "="
-    if base == 0:
-        return f"{head - base:+g}"
-    return f"{(head - base) / abs(base) * 100.0:+.0f}%"
-
-
-def _md_metric_rows(base: dict | None, head: dict) -> list[str]:
-    """A `| Metric | Base | Head | Δ |` table body for one result pair.
-
-    Shared by the aggregate roll-up and the per-scenario tables. A base
-    produced before a metric existed (newly added metric on the PR head) has
-    no value to compare against, so both Base and Δ degrade to ``—``."""
-    rows = ["| Metric | Base | Head | Δ |", "|---|---:|---:|---:|"]
-    for key in _REPORT_METRICS:
-        hv: object = head.get(key, "—")
-        if base is not None and key in base and key in head:
-            bv: object = base[key]
-            delta = _fmt_delta(float(base[key]), float(head[key]))
-        else:
-            bv = "—"
-            delta = "—"
-        rows.append(f"| {key} | {bv} | {hv} | {delta} |")
-    return rows
-
-
-def _seeds_phrase(results: list[dict]) -> str:
-    """`mean of N seeds` / `single seed` describing how a result set was run."""
-    n = max((int(r.get("n_seeds", 1)) for r in results), default=1)
-    return f"mean of {n} seeds" if n > 1 else "single seed"
-
-
-def _seeds_caption(base: list[dict] | None, head: list[dict]) -> str:
-    """One sentence on the seed count behind each side, or ``""`` when both ran a
-    single seed (so the note only appears once figures are seed-averaged)."""
-    hp = _seeds_phrase(head)
-    bp = _seeds_phrase(base) if base else None
-    if hp == "single seed" and (bp is None or bp == "single seed"):
-        return ""
-    if bp is None or bp == hp:
-        return f"Metrics are the per-scenario {hp}."
-    return f"Metrics are the per-scenario mean over seeds (base: {bp}, head: {hp})."
-
-
-def _seeds_note(base: list[dict], head: list[dict]) -> list[str]:
-    """The :func:`_seeds_caption` sentence as italic Markdown lines (empty when
-    there's nothing to note)."""
-    caption = _seeds_caption(base, head)
-    return [f"_{caption}_", ""] if caption else []
-
-
-def render_markdown_compare(
-    base: list[dict], head: list[dict], *, report_available: bool = False
-) -> str:
-    """Markdown before/after tables for the CI PR comment.
-
-    The comment carries the metrics tables for an at-a-glance read; the
-    interactive grid-power charts live in the self-contained HTML report
-    (:func:`astrameter.simulator.eval_report.render_html_report`) that CI
-    uploads as the ``steering-eval`` artifact, since GitHub can't render an
-    interactive chart inline in a comment.
-
-    Set *report_available* when an HTML report is being produced (and CI will
-    append a link to it); only then is the "see the link below" pointer
-    included, so a plain ``--compare`` run doesn't promise a report that
-    doesn't exist.
-    """
-    base_by = {r["scenario"]: r for r in base}
-    base_agg, head_agg = _compare_aggregates(base, head)
-    out = [
-        "### Steering evaluation (base vs head)",
-        "",
-    ]
-    # Lead with the aggregate so a reviewer sees the overall direction before
-    # any per-scenario table — the whole point of the roll-up.
-    if base_agg is not None:
-        out.append(f"**Overall: {_overall_summary(base_agg, head_agg)}.**")
-        out.append("")
-        out.append(f"**Priority: {_priority_summary(base_agg, head_agg)}.**")
-        out.append("")
-    out += [
-        "Lower is better for every metric. See "
-        "`src/astrameter/simulator/evaluation.py` for definitions.",
-        "",
-    ]
-    out += _seeds_note(base, head)
-    out += [
-        f"#### Aggregate — mean across {head_agg['n_scenarios']} scenarios",
-        "",
-    ]
-    out += _md_metric_rows(base_agg, head_agg)
-    out.append("")
-    if report_available:
-        out += [
-            "📊 **Interactive grid-power charts** (zoom / hover / toggle series) "
-            "are in the self-contained `steering-eval-report.html` report — see "
-            "the link below (it opens directly in the browser).",
-            "",
-        ]
-    out += [
-        "<details><summary><b>What do these metrics mean?</b></summary>",
-        "",
-        "| Metric | Meaning |",
-        "|---|---|",
-    ]
-    out.extend(f"| `{key}` | {desc} |" for key, desc in _METRIC_GLOSSARY)
-    out.append("")
-    out.append("</details>")
-    out.append("")
-    # Collapse every per-scenario table behind one outer section so the comment
-    # leads with the aggregate roll-up and verdicts; reviewers expand this only
-    # to drill into individual scenarios (each still its own nested collapsible).
-    out.append(
-        f"<details><summary><b>Per-scenario tables</b> "
-        f"({len(head)} scenarios)</summary>"
-    )
-    out.append("")
-    for res in head:
-        b = base_by.get(res["scenario"])
-        out.append(
-            f"<details><summary><b>{res['scenario']}</b> — "
-            f"{_summary_line(b, res)}</summary>"
-        )
-        out.append("")
-        out.extend(_md_metric_rows(b, res))
-        out.append("")
-        out.append("</details>")
-    out.append("")
-    out.append("</details>")
-    missing = [
-        r["scenario"]
-        for r in base
-        if r["scenario"] not in {h["scenario"] for h in head}
-    ]
-    if missing:
-        out.append("")
-        out.append(f"_Scenarios only in base: {', '.join(missing)}_")
-    return "\n".join(out)
-
-
-def _summary_line(base: dict | None, head: dict) -> str:
-    parts = [
-        f"settle {head['settle_mean_s']}s",
-        f"overshoot {head['overshoot_max_w']}W",
-        f"RMS {head['steady_rms_w']}W",
-    ]
-    if base:
-        parts = [
-            f"settle {base['settle_mean_s']}→{head['settle_mean_s']}s",
-            f"overshoot {base['overshoot_max_w']}→{head['overshoot_max_w']}W",
-            f"RMS {base['steady_rms_w']}→{head['steady_rms_w']}W",
-        ]
-    return ", ".join(parts)
-
-
-# ---------------------------------------------------------------------------
-# CLI
-# ---------------------------------------------------------------------------
 
 
 def _parse_overrides(pairs: list[str]) -> dict[str, float]:
@@ -2344,40 +93,6 @@ def _parse_overrides(pairs: list[str]) -> dict[str, float]:
             raise SystemExit(f"--set expects KEY=VALUE, got {pair!r}")
         overrides[key.strip()] = float(value)
     return overrides
-
-
-def _run_one(name: str, seed: int, overrides: dict[str, float]) -> dict:
-    """Run one (scenario, seed) to completion in its own process.
-
-    A process-pool worker: it rebuilds the scenario registry from *name*
-    (closures in ``Scenario.build_events`` aren't picklable, so we ship the
-    name, not the object) and drives the async scenario under a fresh event
-    loop. Process isolation also gives each seed its own global ``random``
-    state, so concurrent runs stay bit-for-bit deterministic — which a single
-    interpreter (asyncio interleaving the shared global RNG) could not."""
-    scenarios = build_scenarios()
-    return asyncio.run(run_scenario(scenarios[name], seed=seed, overrides=overrides))
-
-
-def _run_tasks(
-    tasks: list[tuple[str, int]], overrides: dict[str, float]
-) -> dict[tuple[str, int], dict]:
-    """Run every (scenario, seed) task, in parallel across CPU cores.
-
-    The simulation is CPU-bound, so processes (not asyncio) are what actually
-    parallelize it; one task runs inline to avoid pool overhead."""
-    if len(tasks) <= 1:
-        return {t: _run_one(t[0], t[1], overrides) for t in tasks}
-    workers = min(len(tasks), os.cpu_count() or 1)
-    out: dict[tuple[str, int], dict] = {}
-    with ProcessPoolExecutor(max_workers=workers) as pool:
-        futures = {
-            pool.submit(_run_one, name, seed, overrides): (name, seed)
-            for name, seed in tasks
-        }
-        for fut in as_completed(futures):
-            out[futures[fut]] = fut.result()
-    return out
 
 
 def _run_all(
@@ -2483,8 +198,6 @@ def main(argv: list[str] | None = None) -> None:
         print(render_text(results))
 
     if args.html:
-        from .eval_report import render_html_report
-
         base_agg, head_agg = _compare_aggregates(base, results)
         report = render_html_report(
             base,
